% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Addressable}
\alias{Addressable}
\title{Addressable Object}
\usage{
Addressable(url = NULL)
}
\arguments{
\item{url}{No description}
}
\value{
Addressable object
}
\description{
Addressable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information for connecting over HTTP(s).
}
\concept{Addressable functions}
