% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{AppEngineHttpTarget}
\alias{AppEngineHttpTarget}
\title{AppEngineHttpTarget Object}
\usage{
AppEngineHttpTarget(AppEngineHttpTarget.headers = NULL,
  relativeUri = NULL, headers = NULL, appEngineRouting = NULL,
  httpMethod = NULL, body = NULL)
}
\arguments{
\item{AppEngineHttpTarget.headers}{The \link{AppEngineHttpTarget.headers} object or list of objects}

\item{relativeUri}{The relative URI}

\item{headers}{HTTP request headers}

\item{appEngineRouting}{App Engine Routing setting for the job}

\item{httpMethod}{The HTTP method to use for the request}

\item{body}{Body}
}
\value{
AppEngineHttpTarget object
}
\description{
AppEngineHttpTarget Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
App Engine target. The job will be pushed to a job handler by meansof an HTTP request via an http_method suchas HTTP POST, HTTP GET, etc. The job is acknowledged by means of anHTTP response code in the range [200 - 299]. Error 503 isconsidered an App Engine system error instead of an applicationerror. Requests returning error 503 will be retried regardless ofretry configuration and not counted against retry counts. Any otherresponse code, or a failure to receive a response before thedeadline, constitutes a failed attempt.
}
\seealso{
Other AppEngineHttpTarget functions: \code{\link{AppEngineHttpTarget.headers}}
}
\concept{AppEngineHttpTarget functions}
