% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{AppEngineRouting}
\alias{AppEngineRouting}
\title{Cloud Scheduler API Objects
Creates and manages jobs run on a regular recurring schedule.}
\usage{
AppEngineRouting(version = NULL, host = NULL, service = NULL,
  instance = NULL)
}
\arguments{
\item{version}{App version}

\item{host}{Output only}

\item{service}{App service}

\item{instance}{App instance}
}
\value{
AppEngineRouting object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2019-11-04 15:44:38
filename: R/objs_schedule.R
api_json: schedule
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AppEngineRouting Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
App Engine Routing.For more information about services, versions, and instances see[An Overview of AppEngine](https://cloud.google.com/appengine/docs/python/an-overview-of-app-engine),[Microservices Architecture on Google AppEngine](https://cloud.google.com/appengine/docs/python/microservices-on-app-engine),[App Engine Standard requestrouting](https://cloud.google.com/appengine/docs/standard/python/how-requests-are-routed),and [App Engine Flex requestrouting](https://cloud.google.com/appengine/docs/flexible/python/how-requests-are-routed).
}
\concept{AppEngineRouting functions}
