% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{ArtifactObjects}
\alias{ArtifactObjects}
\title{ArtifactObjects Object}
\usage{
ArtifactObjects(location = NULL, paths = NULL, timing = NULL)
}
\arguments{
\item{location}{Cloud Storage bucket and optional object path, in the form}

\item{paths}{Path globs used to match files in the build's workspace}

\item{timing}{Output only}
}
\value{
ArtifactObjects object
}
\description{
ArtifactObjects Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Files in the workspace to upload to Cloud Storage upon successfulcompletion of all build steps.
}
\concept{ArtifactObjects functions}
