% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{ArtifactResult}
\alias{ArtifactResult}
\title{ArtifactResult Object}
\usage{
ArtifactResult(fileHash = NULL, location = NULL)
}
\arguments{
\item{fileHash}{The file hash of the artifact}

\item{location}{The path of an artifact in a Google Cloud Storage bucket, with the}
}
\value{
ArtifactResult object
}
\description{
ArtifactResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An artifact that was uploaded during a build. Thisis a single record in the artifact manifest JSON file.
}
\concept{ArtifactResult functions}
