% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{Artifacts}
\alias{Artifacts}
\title{Artifacts Object}
\usage{
Artifacts(images = NULL, objects = NULL)
}
\arguments{
\item{images}{A list of images to be pushed upon the successful completion of all build}

\item{objects}{A list of objects to be uploaded to Cloud Storage upon successful}
}
\value{
Artifacts object
}
\description{
Artifacts Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Artifacts produced by a build that should be uploaded uponsuccessful completion of all build steps.
}
\concept{Artifacts functions}
