% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{AuditConfig}
\alias{AuditConfig}
\title{AuditConfig Object}
\usage{
AuditConfig(service = NULL, auditLogConfigs = NULL)
}
\arguments{
\item{service}{Specifies a service that will be enabled for audit logging}

\item{auditLogConfigs}{The configuration for logging of each type of permission}
}
\value{
AuditConfig object
}
\description{
AuditConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the audit configuration for a service.The configuration determines which permission types are logged, and whatidentities, if any, are exempted from logging.An AuditConfig must have one or more AuditLogConfigs.If there are AuditConfigs for both `allServices` and a specific service,the union of the two AuditConfigs is used for that service: the log_typesspecified in each AuditConfig are enabled, and the exempted_members in eachAuditLogConfig are exempted.Example Policy with multiple AuditConfigs:    {      'audit_configs': [        {          'service': 'allServices'          'audit_log_configs': [            {              'log_type': 'DATA_READ',              'exempted_members': [                'user:jose@example.com'              ]            },            {              'log_type': 'DATA_WRITE',            },            {              'log_type': 'ADMIN_READ',            }          ]        },        {          'service': 'sampleservice.googleapis.com'          'audit_log_configs': [            {              'log_type': 'DATA_READ',            },            {              'log_type': 'DATA_WRITE',              'exempted_members': [                'user:aliya@example.com'              ]            }          ]        }      ]    }For sampleservice, this policy enables DATA_READ, DATA_WRITE and ADMIN_READlogging. It also exempts jose@example.com from DATA_READ logging, andaliya@example.com from DATA_WRITE logging.
}
\concept{AuditConfig functions}
