% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{AuthorizedDomain}
\alias{AuthorizedDomain}
\title{AuthorizedDomain Object}
\usage{
AuthorizedDomain(name = NULL, id = NULL)
}
\arguments{
\item{name}{Read only}

\item{id}{Relative name of the domain authorized for use}
}
\value{
AuthorizedDomain object
}
\description{
AuthorizedDomain Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A domain that a user has been authorized to administer. To authorize useof a domain, verify ownership via[Webmaster Central](https://www.google.com/webmasters/verification/home).
}
\concept{AuthorizedDomain functions}
