% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{AutoDomainMapping}
\alias{AutoDomainMapping}
\title{AutoDomainMapping Object}
\usage{
AutoDomainMapping(metadata = NULL, status = NULL, apiVersion = NULL,
  spec = NULL)
}
\arguments{
\item{metadata}{Metadata associated with this BuildTemplate}

\item{status}{The current status of the AutoDomainMapping}

\item{apiVersion}{The API version for this call such as 'serving}

\item{spec}{The spec for this AutoDomainMapping}
}
\value{
AutoDomainMapping object
}
\description{
AutoDomainMapping Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Resource to hold the state and status of a user's auto domain mapping.
}
\seealso{
Other AutoDomainMapping functions: \code{\link{namespaces.autodomainmappings.create}},
  \code{\link{namespaces.autodomainmappings.replaceAutoDomainMapping}},
  \code{\link{projects.locations.autodomainmappings.create}},
  \code{\link{projects.locations.autodomainmappings.replaceAutoDomainMapping}}
}
\concept{AutoDomainMapping functions}
