% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{AutoDomainMappingSpec}
\alias{AutoDomainMappingSpec}
\title{AutoDomainMappingSpec Object}
\usage{
AutoDomainMappingSpec(forceOverride = NULL, expansionType = NULL,
  certificateMode = NULL)
}
\arguments{
\item{forceOverride}{If set, the mapping will override any mapping set before this spec was set}

\item{expansionType}{The type of expansion for the auto auto domain mapping}

\item{certificateMode}{The mode of the certificate}
}
\value{
AutoDomainMappingSpec object
}
\description{
AutoDomainMappingSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The desired state of the Auto Domain Mapping.
}
\concept{AutoDomainMappingSpec functions}
