% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{AutoDomainMappingStatus}
\alias{AutoDomainMappingStatus}
\title{AutoDomainMappingStatus Object}
\usage{
AutoDomainMappingStatus(conditions = NULL, observedGeneration = NULL,
  resourceRecords = NULL)
}
\arguments{
\item{conditions}{Array of observed AutoDomainMappingConditions, indicating the current state}

\item{observedGeneration}{ObservedGeneration is the 'Generation' of the AutoDomainMapping that}

\item{resourceRecords}{The resource records required to configure this domain mapping}
}
\value{
AutoDomainMappingStatus object
}
\description{
AutoDomainMappingStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The current state of the Domain Mapping.
}
\concept{AutoDomainMappingStatus functions}
