% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{Binding}
\alias{Binding}
\title{Binding Object}
\usage{
Binding(condition = NULL, members = NULL, role = NULL)
}
\arguments{
\item{condition}{The condition that is associated with this binding}

\item{members}{Specifies the identities requesting access for a Cloud Platform resource}

\item{role}{Role that is assigned to `members`}
}
\value{
Binding object
}
\description{
Binding Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Associates `members` with a `role`.
}
\concept{Binding functions}
