% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild.R
\name{Build}
\alias{Build}
\title{Build Object}
\usage{
Build(Build.substitutions = NULL, Build.timing = NULL,
  results = NULL, logsBucket = NULL, steps = NULL,
  buildTriggerId = NULL, id = NULL, tags = NULL, startTime = NULL,
  substitutions = NULL, timing = NULL, sourceProvenance = NULL,
  createTime = NULL, images = NULL, projectId = NULL,
  logUrl = NULL, finishTime = NULL, source = NULL, options = NULL,
  timeout = NULL, status = NULL, statusDetail = NULL,
  artifacts = NULL, secrets = NULL)
}
\arguments{
\item{Build.substitutions}{The \link{Build.substitutions} object or list of objects}

\item{Build.timing}{The \link{Build.timing} object or list of objects}

\item{results}{Output only}

\item{logsBucket}{Google Cloud Storage bucket where logs should be written (see}

\item{steps}{Required}

\item{buildTriggerId}{Output only}

\item{id}{Output only}

\item{tags}{Tags for annotation of a `Build`}

\item{startTime}{Output only}

\item{substitutions}{Substitutions data for `Build` resource}

\item{timing}{Output only}

\item{sourceProvenance}{Output only}

\item{createTime}{Output only}

\item{images}{A list of images to be pushed upon the successful completion of all build}

\item{projectId}{Output only}

\item{logUrl}{Output only}

\item{finishTime}{Output only}

\item{source}{A \link{Source} object specifying the location of the source files to build}

\item{options}{Special options for this build}

\item{timeout}{Amount of time that this build should be allowed to run, to second}

\item{status}{Output only}

\item{statusDetail}{Output only}

\item{artifacts}{Artifacts produced by the build that should be uploaded upon}

\item{secrets}{Secrets to decrypt using Cloud Key Management Service}
}
\value{
Build object
}
\description{
Build Object
}
\details{
A build resource in the Cloud Build API.At a high level, a `Build` describes where to find source code, how to buildit (for example, the builder image to run on the source), and where to storethe built artifacts.
Fields can include the following variables, which will be expanded when the build is created:- $PROJECT_ID: the project ID of the build.- $BUILD_ID: the autogenerated ID of the build.- $REPO_NAME: the source repository name specified by RepoSource.- $BRANCH_NAME: the branch name specified by RepoSource.- $TAG_NAME: the tag name specified by RepoSource.- $REVISION_ID or $COMMIT_SHA: the commit SHA specified by RepoSource or  resolved from the specified branch or tag.- $SHORT_SHA: first 7 characters of $REVISION_ID or $COMMIT_SHA.
}
\seealso{
Other Build functions: \code{\link{cr_build}},
  \code{\link{projects.builds.create}}
}
\concept{Build functions}
