% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{Build.substitutions}
\alias{Build.substitutions}
\title{Build.substitutions Object}
\usage{
Build.substitutions()
}
\value{
Build.substitutions object
}
\description{
Build.substitutions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Substitutions data for `Build` resource.
}
\seealso{
Other Build functions: \code{\link{Build.timing}},
  \code{\link{Build}}, \code{\link{cr_build}},
  \code{\link{projects.builds.create}}
}
\concept{Build functions}
