% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{BuildOptions}
\alias{BuildOptions}
\title{BuildOptions Object}
\usage{
BuildOptions(logStreamingOption = NULL, workerPool = NULL,
  env = NULL, logging = NULL, requestedVerifyOption = NULL,
  substitutionOption = NULL, diskSizeGb = NULL, secretEnv = NULL,
  sourceProvenanceHash = NULL, machineType = NULL, volumes = NULL)
}
\arguments{
\item{logStreamingOption}{Option to define build log streaming behavior to Google Cloud}

\item{workerPool}{Option to specify a `WorkerPool` for the build}

\item{env}{A list of global environment variable definitions that will exist for all}

\item{logging}{Option to specify the logging mode, which determines where the logs are}

\item{requestedVerifyOption}{Requested verifiability options}

\item{substitutionOption}{Option to specify behavior when there is an error in the substitution}

\item{diskSizeGb}{Requested disk size for the VM that runs the build}

\item{secretEnv}{A list of global environment variables, which are encrypted using a Cloud}

\item{sourceProvenanceHash}{Requested hash for SourceProvenance}

\item{machineType}{Compute Engine machine type on which to run the build}

\item{volumes}{Global list of volumes to mount for ALL build steps}
}
\value{
BuildOptions object
}
\description{
BuildOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional arguments to enable specific features of builds.
}
\concept{BuildOptions functions}
