% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{BuildStep}
\alias{BuildStep}
\title{BuildStep Object}
\usage{
BuildStep(args = NULL, pullTiming = NULL, timeout = NULL,
  status = NULL, name = NULL, id = NULL, secretEnv = NULL,
  dir = NULL, env = NULL, waitFor = NULL, timing = NULL,
  entrypoint = NULL, volumes = NULL)
}
\arguments{
\item{args}{A list of arguments that will be presented to the step when it is started}

\item{pullTiming}{Output only}

\item{timeout}{Time limit for executing this build step}

\item{status}{Output only}

\item{name}{Required}

\item{id}{Unique identifier for this build step, used in `wait_for` to}

\item{secretEnv}{A list of environment variables which are encrypted using a Cloud Key}

\item{dir}{Working directory to use when running this step's container}

\item{env}{A list of environment variable definitions to be used when running a step}

\item{waitFor}{The ID(s) of the step(s) that this build step depends on}

\item{timing}{Output only}

\item{entrypoint}{Entrypoint to be used instead of the build step image's default entrypoint}

\item{volumes}{List of volumes to mount into the build step}
}
\value{
BuildStep object
}
\description{
BuildStep Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A step in the build pipeline.
}
\concept{BuildStep functions}
