% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{BuildTrigger}
\alias{BuildTrigger}
\title{BuildTrigger Object}
\usage{
BuildTrigger(BuildTrigger.substitutions = NULL, filename = NULL,
  name = NULL, id = NULL, tags = NULL, build = NULL,
  ignoredFiles = NULL, github = NULL, substitutions = NULL,
  includedFiles = NULL, disabled = NULL, createTime = NULL,
  triggerTemplate = NULL, description = NULL)
}
\arguments{
\item{BuildTrigger.substitutions}{The \link{BuildTrigger.substitutions} object or list of objects}

\item{filename}{Path, from the source root, to a file whose contents is used for the}

\item{name}{User assigned name of the trigger}

\item{id}{Output only}

\item{tags}{Tags for annotation of a `BuildTrigger`}

\item{build}{Contents of the build template}

\item{ignoredFiles}{ignored_files and included_files are file glob matches using}

\item{github}{GitHubEventsConfig describes the configuration of a trigger that creates}

\item{substitutions}{Substitutions data for Build resource}

\item{includedFiles}{If any of the files altered in the commit pass the ignored_files}

\item{disabled}{If true, the trigger will never result in a build}

\item{createTime}{Output only}

\item{triggerTemplate}{Template describing the types of source changes to trigger a build}

\item{description}{Human-readable description of this trigger}
}
\value{
BuildTrigger object
}
\description{
BuildTrigger Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for an automated build in response to source repositorychanges.
}
\seealso{
Other BuildTrigger functions: \code{\link{cr_buildtrigger_edit}},
  \code{\link{cr_buildtrigger}}
}
\concept{BuildTrigger functions}
