% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{BuildTrigger.substitutions}
\alias{BuildTrigger.substitutions}
\title{BuildTrigger.substitutions Object}
\usage{
BuildTrigger.substitutions()
}
\value{
BuildTrigger.substitutions object
}
\description{
BuildTrigger.substitutions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Substitutions data for Build resource.
}
\seealso{
Other BuildTrigger functions: \code{\link{BuildTrigger}},
  \code{\link{projects.triggers.create}},
  \code{\link{projects.triggers.patch}}
}
\concept{BuildTrigger functions}
