% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{BuiltImage}
\alias{BuiltImage}
\title{BuiltImage Object}
\usage{
BuiltImage(name = NULL, digest = NULL, pushTiming = NULL)
}
\arguments{
\item{name}{Name used to push the container image to Google Container Registry, as}

\item{digest}{Docker Registry 2}

\item{pushTiming}{Output only}
}
\value{
BuiltImage object
}
\description{
BuiltImage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An image built by the pipeline.
}
\concept{BuiltImage functions}
