% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{CancelBuildRequest}
\alias{CancelBuildRequest}
\title{CancelBuildRequest Object}
\usage{
CancelBuildRequest()
}
\value{
CancelBuildRequest object
}
\description{
CancelBuildRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to cancel an ongoing build.
}
\seealso{
Other CancelBuildRequest functions: \code{\link{projects.builds.cancel}}
}
\concept{CancelBuildRequest functions}
