% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{CancelOperationRequest}
\alias{CancelOperationRequest}
\title{CancelOperationRequest Object}
\usage{
CancelOperationRequest()
}
\value{
CancelOperationRequest object
}
\description{
CancelOperationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for Operations.CancelOperation.
}
\seealso{
Other CancelOperationRequest functions: \code{\link{operations.cancel}}
}
\concept{CancelOperationRequest functions}
