% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_build.R
\docType{package}
\name{cloudbuild_googleAuthR}
\alias{cloudbuild_googleAuthR}
\alias{cloudbuild_googleAuthR-package}
\title{Cloud Build API
Creates and manages builds on Google Cloud Platform.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2019-11-03 21:21:25
filename: R/funcs_build.R
api_json: build
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
