% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudRunner.R
\docType{package}
\name{cloudRunner}
\alias{cloudRunner}
\alias{cloudRunner-package}
\title{Launch R scripts into the Google Cloud via Cloud Build, Cloud Run and Cloud Scheduler}
\description{
Launch R scripts into the Google Cloud via Cloud Build, Cloud Run and Cloud Scheduler
}
