% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_schedule.R
\docType{package}
\name{cloudscheduler_googleAuthR}
\alias{cloudscheduler_googleAuthR}
\alias{cloudscheduler_googleAuthR-package}
\title{Cloud Scheduler API
Creates and manages jobs run on a regular recurring schedule.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2019-11-04 15:44:16
filename: R/funcs_schedule.R
api_json: schedule
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
