% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ConfigMapEnvSource}
\alias{ConfigMapEnvSource}
\title{ConfigMapEnvSource Object}
\usage{
ConfigMapEnvSource(localObjectReference = NULL, optional = NULL,
  name = NULL)
}
\arguments{
\item{localObjectReference}{This field should not be used directly as it is meant to be inlined}

\item{optional}{(Optional)}

\item{name}{Cloud Run fully managed: not supported}
}
\value{
ConfigMapEnvSource object
}
\description{
ConfigMapEnvSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedConfigMapEnvSource selects a ConfigMap to populate the environmentvariables with.The contents of the target ConfigMap's Data field will represent thekey-value pairs as environment variables.
}
\concept{ConfigMapEnvSource functions}
