% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ConfigMapKeySelector}
\alias{ConfigMapKeySelector}
\title{ConfigMapKeySelector Object}
\usage{
ConfigMapKeySelector(optional = NULL, key = NULL, name = NULL,
  localObjectReference = NULL)
}
\arguments{
\item{optional}{(Optional)}

\item{key}{Cloud Run fully managed: not supported}

\item{name}{Cloud Run fully managed: not supported}

\item{localObjectReference}{This field should not be used directly as it is meant to be inlined}
}
\value{
ConfigMapKeySelector object
}
\description{
ConfigMapKeySelector Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedSelects a key from a ConfigMap.
}
\concept{ConfigMapKeySelector functions}
