% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ConfigMapVolumeSource}
\alias{ConfigMapVolumeSource}
\title{ConfigMapVolumeSource Object}
\usage{
ConfigMapVolumeSource(optional = NULL, name = NULL,
  defaultMode = NULL, items = NULL)
}
\arguments{
\item{optional}{(Optional)}

\item{name}{Cloud Run fully managed: not supported}

\item{defaultMode}{(Optional)}

\item{items}{(Optional)}
}
\value{
ConfigMapVolumeSource object
}
\description{
ConfigMapVolumeSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedAdapts a ConfigMap into a volume.The contents of the target ConfigMap's Data field will be presented in avolume as files using the keys in the Data field as the file names, unlessthe items element is populated with specific mappings of keys to paths.
}
\concept{ConfigMapVolumeSource functions}
