% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Configuration}
\alias{Configuration}
\title{Configuration Object}
\usage{
Configuration(metadata = NULL, status = NULL, apiVersion = NULL,
  spec = NULL)
}
\arguments{
\item{metadata}{Metadata associated with this Configuration, including name, namespace,}

\item{status}{Status communicates the observed state of the Configuration (from the}

\item{apiVersion}{The API version for this call such as 'serving}

\item{spec}{Spec holds the desired state of the Configuration (from the client)}
}
\value{
Configuration object
}
\description{
Configuration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration represents the 'floating HEAD' of a linear history ofRevisions, and optionally how the containers those revisions reference arebuilt. Users create new Revisions by updating the Configuration's spec. The'latest created' revision's name is available under status, as is the 'latestready' revision's name. See also:https://github.com/knative/serving/blob/master/docs/spec/overview.md#configuration
}
\seealso{
Other Configuration functions: \code{\link{namespaces.configurations.create}},
  \code{\link{namespaces.configurations.replaceConfiguration}},
  \code{\link{projects.locations.configurations.create}},
  \code{\link{projects.locations.configurations.replaceConfiguration}}
}
\concept{Configuration functions}
