% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ConfigurationSpec}
\alias{ConfigurationSpec}
\title{ConfigurationSpec Object}
\usage{
ConfigurationSpec(template = NULL)
}
\arguments{
\item{template}{Optional metadata for this Revision, including labels and annotations}
}
\value{
ConfigurationSpec object
}
\description{
ConfigurationSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ConfigurationSpec holds the desired state of the Configuration (from theclient).
}
\concept{ConfigurationSpec functions}
