% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ConfigurationStatus}
\alias{ConfigurationStatus}
\title{ConfigurationStatus Object}
\usage{
ConfigurationStatus(latestCreatedRevisionName = NULL,
  latestReadyRevisionName = NULL, conditions = NULL,
  observedGeneration = NULL)
}
\arguments{
\item{latestCreatedRevisionName}{LatestCreatedRevisionName is the last revision that was created from this}

\item{latestReadyRevisionName}{LatestReadyRevisionName holds the name of the latest Revision stamped out}

\item{conditions}{Conditions communicates information about ongoing/complete}

\item{observedGeneration}{ObservedGeneration is the 'Generation' of the Configuration that}
}
\value{
ConfigurationStatus object
}
\description{
ConfigurationStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ConfigurationStatus communicates the observed state of the Configuration(from the controller).
}
\concept{ConfigurationStatus functions}
