% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Container}
\alias{Container}
\title{Container Object}
\usage{
Container(volumeMounts = NULL, ports = NULL, args = NULL,
  workingDir = NULL, envFrom = NULL, image = NULL, name = NULL,
  command = NULL, terminationMessagePolicy = NULL,
  securityContext = NULL, livenessProbe = NULL, env = NULL,
  imagePullPolicy = NULL, readinessProbe = NULL, resources = NULL,
  terminationMessagePath = NULL)
}
\arguments{
\item{volumeMounts}{(Optional)}

\item{ports}{(Optional)}

\item{args}{(Optional)}

\item{workingDir}{(Optional)}

\item{envFrom}{(Optional)}

\item{image}{Cloud Run fully managed: only supports containers from Google Container}

\item{name}{(Optional)}

\item{command}{No description}

\item{terminationMessagePolicy}{(Optional)}

\item{securityContext}{(Optional)}

\item{livenessProbe}{(Optional)}

\item{env}{(Optional)}

\item{imagePullPolicy}{(Optional)}

\item{readinessProbe}{(Optional)}

\item{resources}{(Optional)}

\item{terminationMessagePath}{(Optional)}
}
\value{
Container object
}
\description{
Container Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single application container.This specifies both the container to run, the command to run in the containerand the arguments to supply to it.Note that additional arguments may be supplied by the system to the containerat runtime.
}
\concept{Container functions}
