% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ContainerPort}
\alias{ContainerPort}
\title{ContainerPort Object}
\usage{
ContainerPort(name = NULL, containerPort = NULL, protocol = NULL)
}
\arguments{
\item{name}{(Optional)}

\item{containerPort}{(Optional)}

\item{protocol}{(Optional)}
}
\value{
ContainerPort object
}
\description{
ContainerPort Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedContainerPort represents a network port in a single container.
}
\concept{ContainerPort functions}
