% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild.R
\name{cr_build}
\alias{cr_build}
\title{Starts a build with the specified configuration.}
\usage{
cr_build(yaml, timeout = NULL, images = NULL,
  projectId = Sys.getenv("GCE_DEFAULT_PROJECT_ID"))
}
\arguments{
\item{yaml}{A cloudbuild.yaml with the steps to run for the build}

\item{timeout}{Amount of time that this build should be allowed to run, to second}

\item{images}{A list of images to be pushed upon the successful completion of all build}

\item{projectId}{ID of the project}
}
\description{
This method returns a long-running `Operation`, which includes the buildID. Pass the build ID to `GetBuild` to determine the build status (such as`SUCCESS` or `FAILURE`).
}
\seealso{
\href{https://cloud.google.com/cloud-build/docs/}{Google Documentation for Cloud Build}

Other Build functions: \code{\link{Build.substitutions}},
  \code{\link{Build.timing}}, \code{\link{Build}},
  \code{\link{projects.builds.create}}
}
\concept{Build functions}
