% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_admin.R
\name{cr_build_list}
\alias{cr_build_list}
\title{Lists the build}
\usage{
cr_build_list(
  filter = NULL,
  projectId = cr_project_get(),
  pageSize = 1000,
  data_frame_output = TRUE
)
}
\arguments{
\item{filter}{Text filter for the list}

\item{projectId}{ID of the project}

\item{pageSize}{How many builds to fetch per page}

\item{data_frame_output}{If TRUE will output a data.frame of a subset of info from the builds, merged with the list of triggers from \link{cr_buildtrigger_list}.  Set to FALSE to return a list of Build objects similar to output from \link{cr_build_status}}
}
\description{
Get a list of builds within your project
}
\details{
If filter is \code{NULL} then this will return all historic builds which can be a lot.  Its probably better to have a filter, but there is no documentation on the filter syntax at the moment.
}
\examples{

\dontrun{

 # merge with buildtrigger list
 cr_build_list()

 # output a list of build objects
 cr_build_list(data_frame_output=FALSE)
}
}
\seealso{
\url{https://cloud.google.com/cloud-build/docs/api/reference/rest/v1/projects.builds/list}

Other Cloud Build functions: 
\code{\link{Build}()},
\code{\link{RepoSource}()},
\code{\link{Source}()},
\code{\link{StorageSource}()},
\code{\link{cr_build_artifacts}()},
\code{\link{cr_build_logs}()},
\code{\link{cr_build_make}()},
\code{\link{cr_build_status}()},
\code{\link{cr_build_upload_gcs}()},
\code{\link{cr_build_wait}()},
\code{\link{cr_build_write}()},
\code{\link{cr_build_yaml_artifact}()},
\code{\link{cr_build_yaml_secrets}()},
\code{\link{cr_build_yaml}()},
\code{\link{cr_build}()}
}
\concept{Cloud Build functions}
