% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sources.R
\name{cr_build_source}
\alias{cr_build_source}
\alias{cr_build_source.gar_RepoSource}
\alias{cr_build_source.gar_StorageSource}
\title{Build a source object}
\usage{
cr_build_source(x)

\method{cr_build_source}{gar_RepoSource}(x)

\method{cr_build_source}{gar_StorageSource}(x)
}
\arguments{
\item{x}{A \link{RepoSource} or a \link{StorageSource} object}
}
\description{
Build a source object
}
\examples{

cr_build_source(RepoSource("my_repo", branchName = "master"))
cr_build_source(StorageSource("my_code.tar.gz","gs://my-bucket"))

my_gcs_source <- cr_build_source("my_code.tar.gz", "gs://my-bucket"))
my_repo_source <- cr_build_source(RepoSource("https://my-repo.com",
                                             branchName="master"))

build1 <- cr_build("cloudbuild.yaml", source = my_gcs_source)
build2 <- cr_build("cloudbuild.yaml", source = my_repo_source)
}
