% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild.R
\name{cr_build_status}
\alias{cr_build_status}
\title{Returns information about a previously requested build.}
\usage{
cr_build_status(id, projectId = Sys.getenv("GCE_DEFAULT_PROJECT_ID"))
}
\arguments{
\item{id}{ID of the build or a \code{BuildOperationMetadata} object}

\item{projectId}{ID of the project}
}
\description{
The `Build` that is returned includes its status (such as `SUCCESS`,`FAILURE`, or `WORKING`), and timing information.
}
