% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_targets.R
\name{cr_build_tar_target}
\alias{cr_build_tar_target}
\title{Use this within your _targets functions to send the step to Cloud Build}
\usage{
cr_build_tar_target(
  x,
  ...,
  bucket = cr_bucket_get(),
  target_folder = NULL,
  task_args = NULL,
  task_image = task_image
)
}
\arguments{
\item{x}{The tar_target node name (maybe derive this?)}

\item{...}{The arguments to send to the function that will run on Cloud Build}

\item{bucket}{The Google Cloud Storage bucket the target metadata will be saved to in folder `target_folder`}

\item{target_folder}{Where target metadata will sit within the Google Cloud Storage bucket as a folder.  If NULL defaults to RStudio project name or "targets_cloudbuild" if no RStudio project found.}

\item{task_args}{A named list of additional arguments to send to \link{cr_buildstep_r} when its executing the \link[targets]{tar_make} command (such as environment arguments)}

\item{task_image}{An existing Docker image that will be used to run your targets workflow after the targets meta has been downloaded from Google Cloud Storage}
}
\description{
Use this within your _targets functions to send the step to Cloud Build
}
