% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild.R
\name{cr_build_upload_gcs}
\alias{cr_build_upload_gcs}
\title{Create a StorageSource}
\usage{
cr_build_upload_gcs(local, remote = paste0(local, format(Sys.time(),
  "\%Y\%m\%d\%H\%M\%S"), "tar.gz"),
  bucket = Sys.getenv("GCS_DEFAULT_BUCKET"))
}
\arguments{
\item{local}{Local directory containing the Dockerfile etc. you wish to deploy}

\item{remote}{The name of the folder in your bucket}

\item{bucket}{The Google Cloud Storage bucket to uplaod to}
}
\description{
This creates a \link{StorageSource} object after uploading to Google Cloud Storage
}
\examples{

\dontrun{

storage <- cr_build_upload_gcs("my_folder")
my_gcs_source <- Source(storageSource=storage)
build1 <- cr_build("cloudbuild.yaml", source = my_gcs_source)

}
}
