% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild.R
\name{cr_build_wait}
\alias{cr_build_wait}
\title{Wait for a Build to run}
\usage{
cr_build_wait(op = .Last.value,
  projectId = Sys.getenv("GCE_DEFAULT_PROJECT_ID"))
}
\arguments{
\item{op}{The operation build object to wait for}

\item{projectId}{The projectId}
}
\description{
This will repeatedly call \link{cr_build_status} whilst the status is STATUS_UNKNOWN, QUEUED or WORKING
}
