% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{cr_build_yaml}
\alias{cr_build_yaml}
\title{Create a cloudbuild Yaml object in R}
\usage{
cr_build_yaml(steps, timeout = NULL, logsBucket = NULL,
  options = NULL, substitutions = NULL, tags = NULL,
  secrets = NULL, images = NULL, artifacts = NULL)
}
\arguments{
\item{steps}{A vector of \link{cr_buildstep}}

\item{timeout}{How long the entrie build will run. If not set will be 10mins}

\item{logsBucket}{Where logs are written.  If you don't set this field, Cloud Build will use a default bucket to store your build logs.}

\item{options}{A named list of options}

\item{substitutions}{Build macros that will replace entries in other elements}

\item{tags}{Tags for the build}

\item{secrets}{A secrets object}

\item{images}{What images will be build from this cloudbuild}

\item{artificats}{What artifacts may be built from this cloudbuild}
}
\description{
This can be written to disk or used directly with functions such as \link{cr_build}
}
