% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildsteps.R
\name{cr_buildstep_decrypt}
\alias{cr_buildstep_decrypt}
\title{Create a build step for decrypting files via KMS}
\usage{
cr_buildstep_decrypt(cipher, plain, keyring, key, locataion = "global")
}
\arguments{
\item{cipher}{The file that has been encrypted}

\item{plain}{The file location to decrypt to}

\item{keyring}{The KMS keyring to use}

\item{key}{The KMS key to use}

\item{location}{The KMS location}
}
\description{
Create a build step for decrypting files via KMS
}
