% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{cr_buildstep_docker_auth}
\alias{cr_buildstep_docker_auth}
\alias{cr_buildstep_docker_auth_auto}
\title{Authorize Docker using `gcloud auth configure-docker`}
\usage{
cr_buildstep_docker_auth(registry, ...)

cr_buildstep_docker_auth_auto(image, ...)
}
\arguments{
\item{registry}{The registry to authorize `docker`}

\item{...}{Other arguments pass to [googleCloudRunner::cr_buildstep_gcloud()]}

\item{image}{name of the Docker image to push or pull from that needs
authorization}
}
\value{
A buildstep
}
\description{
Authorize Docker using `gcloud auth configure-docker`
}
\examples{
cr_buildstep_docker_auth("us.gcr.io")
cr_buildstep_docker_auth_auto("https://asia.cr.io/myrepo/image")
}
