% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildsteps.R
\name{cr_buildstep_edit}
\alias{cr_buildstep_edit}
\title{Modify an existing buildstep with new parameters}
\usage{
cr_buildstep_edit(x, ...)
}
\arguments{
\item{x}{A buildstep created previously}

\item{...}{Arguments passed on to \code{cr_buildstep}
\describe{
  \item{name}{name of docker image to call appended to \code{prefix}}
  \item{args}{character vector of arguments}
  \item{prefix}{prefixed to name - set to "" to suppress.  Will be suppressed if \code{name} starts with gcr.io}
  \item{entrypoint}{change the entrypoint for the docker container}
  \item{dir}{The directory to use, relative to /workspace e.g. /workspace/deploy/}
  \item{id}{Optional id for the step}
  \item{env}{Environment variables for this step.  A character vector for each assignment}
  \item{volumes}{volumes to connect and write to}
}}
}
\description{
Useful for editing existing buildsteps
}
\examples{
package_build <- system.file("cloudbuild/cloudbuild_packages.yml",
                             package = "googleCloudRunner")
build <- cr_build_make(package_build)
build
cr_buildstep_extract(build, step = 1)
cr_buildstep_extract(build, step = 2)

edit_me <- cr_buildstep_extract(build, step = 2)
cr_buildstep_edit(edit_me, name = "blah")
cr_buildstep_edit(edit_me, name = "gcr.io/blah")
cr_buildstep_edit(edit_me, args = c("blah1","blah2"), dir = "meh")
}
