% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildstep_email.R
\name{cr_buildstep_email}
\alias{cr_buildstep_email}
\title{Send an email in a Cloud Build step}
\usage{
cr_buildstep_email(
  message,
  from,
  to,
  subject,
  cipher = "email_creds.enc",
  plain = "email_creds",
  keyring = "my_keyring",
  key = "email_creds",
  location = "global",
  ...
)
}
\arguments{
\item{message}{The message markdown}

\item{from}{from email}

\item{to}{to email}

\item{subject}{subject email}

\item{cipher}{The file that has been encrypted}

\item{plain}{The file location to decrypt to}

\item{keyring}{The KMS keyring to use}

\item{key}{The KMS key to use}

\item{location}{The KMS location}

\item{...}{Further arguments passed in to \link{cr_buildstep}}
}
\description{
This uses \link[blastula]{smtp_send} to send markdown emails.
}
\details{
You will need to create a credentials file locally first, then deploy it to Key Management Store.  This will be used to send the emails securely.
}
