% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildsteps_templates.R
\name{cr_buildstep_gitsetup}
\alias{cr_buildstep_gitsetup}
\alias{cr_buildstep_git}
\title{Create a build step for authenticating with Git}
\usage{
cr_buildstep_gitsetup(keyring, key, cipher = "id_rsa.enc", ...)

cr_buildstep_git(git_args = c("clone",
  "git@github.com:[GIT-USERNAME]/[REPOSITORY]"), ...)
}
\arguments{
\item{keyring}{The KMS keyring to use}

\item{key}{The KMS key to use}

\item{cipher}{The file that has been encrypted}

\item{...}{Further arguments passed in to \link{cr_buildstep}}

\item{git_args}{The arguments to send to git}
}
\description{
This creates steps to configure git to use an ssh created key.

This creates steps to use git with an ssh created key.
}
\details{
The key should be encrypted offline using \code{gcloud kms} or similar first.  See \link{cr_buildstep_decrypt} for details.

By default the encrypted key should then be at the root of your \link{Source} object called "id_rsa.enc"

You also need to supply a \code{known_hosts} file in working directory.  This can be created by \code{ssh-keyscan -t rsa github.com > known_hosts}

\code{cr_buildstep} must come after \code{cr_buildstep_gitsetup}
}
\examples{

# assumes you have previously saved git ssh key via KMS called "git_key"
Yaml(
     steps = c(
          cr_buildstep_gitsetup("my_keyring", "git_key"),
          cr_buildstep_git(c("clone", "git@github.com:github_name/repo_name"))
     )
 )


}
