% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildsteps_templates.R
\name{cr_buildstep_gitsetup}
\alias{cr_buildstep_gitsetup}
\title{Create a build step for authenticating with Git}
\usage{
cr_buildstep_gitsetup(...)
}
\arguments{
\item{...}{Further arguments passed in to \link{cr_buildstep}}
}
\description{
This creates steps to configure git to use an ssh created key.
}
\details{
The key should be encrypted offline using \code{gcloud kms} or similar first, which you can then use \link{cr_buildstep_decrypt} to put into your builds workspace.
}
\examples{

}
