% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildsteps_templates.R
\name{cr_buildstep_packagetests}
\alias{cr_buildstep_packagetests}
\title{Do R package tests and upload to Codecov}
\usage{
cr_buildstep_packagetests(
  test_script = NULL,
  codecov_script = NULL,
  codecov_token = "$_CODECOV_TOKEN",
  build_image = "gcr.io/gcer-public/packagetools:master",
  env = c("NOT_CRAN=true")
)
}
\arguments{
\item{test_script}{The script that will run first making tests.  If \code{NULL} a default script it used}

\item{codecov_script}{The script that will run first making tests.  If \code{NULL} a default script it used}

\item{codecov_token}{If using codecov, supply your codecov token here. Default assumes you add it to the Cloud Build substitution macros, which is more secure and recommended.}

\item{build_image}{The docker image that will be used to run the R code for the test scripts}

\item{env}{Environment arguments to be set during the test script runs}
}
\description{
This lets you run R package tests and is intended to be used in a trigger when you push to a repository so you can monitor code quality.
}
\examples{

cr_buildstep_packagetests()

}
