% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger_delete}
\alias{cr_buildtrigger_delete}
\title{Deletes a `BuildTrigger` by its project ID and trigger ID.This API is experimental.}
\usage{
cr_buildtrigger_delete(projectId, triggerId)
}
\arguments{
\item{projectId}{ID of the project that owns the trigger}

\item{triggerId}{ID of the `BuildTrigger` to delete}
}
\description{
Deletes a `BuildTrigger` by its project ID and trigger ID.This API is experimental.
}
