% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger_get}
\alias{cr_buildtrigger_get}
\title{Returns information about a `BuildTrigger`.This API is experimental.}
\usage{
cr_buildtrigger_get(projectId, triggerId)
}
\arguments{
\item{projectId}{ID of the project that owns the trigger}

\item{triggerId}{ID of the `BuildTrigger` to get}
}
\description{
Returns information about a `BuildTrigger`.This API is experimental.
}
