% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger_list}
\alias{cr_buildtrigger_list}
\title{Lists existing `BuildTrigger`s.This API is experimental.}
\usage{
cr_buildtrigger_list(projectId, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{projectId}{ID of the project for which to list BuildTriggers}

\item{pageToken}{Token to provide to skip to a particular spot in the list}

\item{pageSize}{Number of results to return in the list}
}
\description{
Lists existing `BuildTrigger`s.This API is experimental.
}
