% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers_events.R
\name{cr_buildtrigger_pubsub}
\alias{cr_buildtrigger_pubsub}
\title{Create a buildtrigger pub/sub object}
\usage{
cr_buildtrigger_pubsub(
  topic,
  serviceAccountEmail = NULL,
  projectId = cr_project_get()
)
}
\arguments{
\item{topic}{The name of the topic from which this subscription is receiving messages.}

\item{serviceAccountEmail}{Service account that will make the push request.}

\item{projectId}{The GCP project the topic is created within}
}
\description{
Create a trigger from a Pub/Sub topic
}
\seealso{
Other BuildTrigger functions: 
\code{\link{BuildTrigger}()},
\code{\link{GitHubEventsConfig}()},
\code{\link{cr_buildtrigger_copy}()},
\code{\link{cr_buildtrigger_delete}()},
\code{\link{cr_buildtrigger_edit}()},
\code{\link{cr_buildtrigger_get}()},
\code{\link{cr_buildtrigger_list}()},
\code{\link{cr_buildtrigger_repo}()},
\code{\link{cr_buildtrigger_run}()},
\code{\link{cr_buildtrigger_webhook}()},
\code{\link{cr_buildtrigger}()}
}
\concept{BuildTrigger functions}
