% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger_run}
\alias{cr_buildtrigger_run}
\title{Runs a `BuildTrigger` at a particular source revision.}
\usage{
cr_buildtrigger_run(RepoSource, projectId, triggerId)
}
\arguments{
\item{RepoSource}{The \link{RepoSource} object to pass to this method}

\item{projectId}{ID of the project}

\item{triggerId}{ID of the trigger}
}
\description{
Runs a `BuildTrigger` at a particular source revision.
}
\seealso{
Other BuildTrigger functions: \code{\link{BuildTrigger}},
  \code{\link{GitHubEventsConfig}},
  \code{\link{cr_buildtrigger_delete}},
  \code{\link{cr_buildtrigger_edit}},
  \code{\link{cr_buildtrigger_get}},
  \code{\link{cr_buildtrigger_list}},
  \code{\link{cr_buildtrigger}}
}
\concept{BuildTrigger functions}
