% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers_events.R
\name{cr_buildtrigger_webhook}
\alias{cr_buildtrigger_webhook}
\title{Create a buildtrigger webhook object}
\usage{
cr_buildtrigger_webhook(secret)
}
\arguments{
\item{secret}{Resource name for the secret required as a URL parameter.}
}
\description{
Create a trigger from a webhook
}
\seealso{
Other BuildTrigger functions: 
\code{\link{BuildTrigger}()},
\code{\link{GitHubEventsConfig}()},
\code{\link{cr_buildtrigger_copy}()},
\code{\link{cr_buildtrigger_delete}()},
\code{\link{cr_buildtrigger_edit}()},
\code{\link{cr_buildtrigger_get}()},
\code{\link{cr_buildtrigger_list}()},
\code{\link{cr_buildtrigger_pubsub}()},
\code{\link{cr_buildtrigger_repo}()},
\code{\link{cr_buildtrigger_run}()},
\code{\link{cr_buildtrigger}()}
}
\concept{BuildTrigger functions}
