% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{cr_deploy}
\alias{cr_deploy}
\title{Deploy an R plumber script to Cloud Run}
\usage{
cr_deploy(local, remote = basename(local), dockerfile = NULL,
  image_name = remote, region = cr_region_get(),
  bucket = cr_bucket_get(), projectId = cr_project_get())
}
\arguments{
\item{local}{A folder containing the R script using plumber called api.R and all its dependencies}

\item{remote}{The folder on Google Cloud Storage}

\item{dockerfile}{An optional Dockerfile built to support the script.  Not needed if 'Dockerfile' exists in folder.  If supplied will be copied into deployment folder and called "Dockerfile"}

\item{image_name}{The gcr.io image name that will be deployed and/or built}

\item{region}{The Cloud Run endpoint set by CR_REGION env arg}

\item{bucket}{The Cloud Storage bucket that will hold the code}

\item{projectId}{The projectId where it all gets deployed to}
}
\description{
Helper to take an R plumber script, create the Dockerfile, add the build to Cloud Build and deploy to Cloud Run
}
\details{
The entrypoint for CloudRun will be via a plumber script called api.R - this should be included in your local folder to deploy.
From that api.R you can source or call other resources in the same folder, using relative paths.

The function will create a local folder called "deploy" and a tar.gz of that folder which is what is being uploaded to Google Cloud Storage
}
\examples{

\dontrun{

cr_deploy(system.file("example/", package = "cloudRunner"))

}
}
