% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{cr_deploy_docker}
\alias{cr_deploy_docker}
\title{Deploy a Dockerfile so it will be built on ContainerRegistry}
\usage{
cr_deploy_docker(local, image_name = remote, dockerfile = NULL,
  remote = basename(local), tag = "$BUILD_ID",
  bucket = cr_bucket_get(), projectId = cr_project_get(),
  launch_browser = interactive())
}
\arguments{
\item{local}{The folder containing the Dockerfile to build}

\item{dockerfile}{An optional Dockerfile built to support the script.  Not needed if 'Dockerfile' exists in folder.  If supplied will be copied into deployment folder and called "Dockerfile"}

\item{remote}{The folder on Google Cloud Storage}

\item{tag}{The tag to attached to the pushed image - can use \code{Build} macros}

\item{bucket}{The GCS bucker that will be used to deploy code source}

\item{projectId}{The projectId}

\item{launch_browser}{Whether to launch the logs URL in a browser once deployed}
}
\description{
If no Dockerfile present in the deployment folder, will attempt to create a Dockerfile to upload via \link{cr_dockerfile}
}
\examples{

\dontrun{

cr_deploy_dockerfile(system.file("example", package="cloudRunner"))

}
}
