% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{cr_deploy_docker_github}
\alias{cr_deploy_docker_github}
\title{Deploy Docker build from a GitHub repo}
\usage{
cr_deploy_docker_github(x, image = x, branch = ".*",
  image_tag = "$SHORT_SHA", dockerfile_location = ".",
  github_tag = NULL, projectId = cr_project_get())
}
\arguments{
\item{x}{The GitHub repo e.g. \code{MarkEdmondson1234/googleCloudRunner}}

\item{image}{The name of the image you want to build}

\item{branch}{A regex of the GitHub branches that will trigger a build}

\item{image_tag}{What to tag the build docker image}

\item{dockerfile_location}{Where the Dockerfile sits within the GitHub repo}

\item{github_tag}{Regexes matching what tags to build. If not NULL then argument branch will be ignored}

\item{projectId}{The project to build under}
}
\description{
This helps the common use case of building a Dockerfile based on the contents of a GitHub repo, and sets up a build trigger so it will build on every commit.
}
\seealso{
\link{cr_deploy_docker} which lets you build Dockerfiles for more generic use cases
}
