% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_gadget.R
\name{cr_deploy_gadget}
\alias{cr_deploy_gadget}
\title{Launch the googleCloudRunner deployment RStudio gadget}
\usage{
cr_deploy_gadget()
}
\description{
You can assign a hotkey to the addin via Tools > Addins > Browse Addins > Keyboard shortcuts.  CTRL+SHIFT+D is a suggested hotkey.
}
