% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy-github.R
\name{cr_deploy_github_html}
\alias{cr_deploy_github_html}
\title{Deploy HTML built from a GitHub repo each commit}
\usage{
cr_deploy_github_html(
  x,
  image = paste0(x, "-html"),
  rmd_folder = NULL,
  html_folder = NULL,
  branch = ".*",
  image_tag = "$SHORT_SHA",
  github_tag = NULL,
  timeout = 600L,
  edit_r = NULL,
  region = cr_region_get(),
  projectId = cr_project_get()
)
}
\arguments{
\item{x}{The GitHub repo e.g. \code{MarkEdmondson1234/googleCloudRunner}}

\item{image}{The name of the image you want to build}

\item{rmd_folder}{A folder of Rmd files within GitHub source that will be built into HTML for serving via \link[rmarkdown]{render}}

\item{html_folder}{A folder of html to deploy within GitHub source.  Will be ignored if rmd_folder is not NULL}

\item{branch}{A regex of the GitHub branches that will trigger a build}

\item{image_tag}{What to tag the build docker image}

\item{github_tag}{Regexes matching what tags to build. If not NULL then argument branch will be ignored}

\item{timeout}{timeout for the Docker build}

\item{edit_r}{If you want to change the R code to render the HTML, supply R code via a file or string of R as per \link{cr_buildstep_r}}

\item{region}{The region for cloud run}

\item{projectId}{The project to build under}
}
\description{
This lets you set up triggers that will update a website each commit
}
\details{
This default R code is rendered in the rmd_folder:

\code{lapply(list.files('.', pattern = '.Rmd', full.names = TRUE),
      rmarkdown::render, output_format = 'html_document')}

You need to mirror the repo onto Google Cloud Repositories, as well as connect to the GitHub app for the source and the build trigger to work from the same GitHub repo.
}
\examples{

\dontrun{
  cr_deploy_github_html("MarkEdmondson1234/googleCloudRunner",
                        rmd_folder = "vignettes")
}
}
\seealso{
\link{cr_deploy_html} that lets you deploy HTML files

Other Deployment functions: 
\code{\link{cr_deploy_docker}()},
\code{\link{cr_deploy_github_docker}()},
\code{\link{cr_deploy_pkgdown}()},
\code{\link{cr_deploy_run}()},
\code{\link{cr_deploy_r}()}
}
\concept{Deployment functions}
