% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy-run.R
\name{cr_deploy_html}
\alias{cr_deploy_html}
\title{Deploy html files to a nginx server on Cloud Run}
\usage{
cr_deploy_html(
  html_folder,
  remote = basename(html_folder),
  image_name = remote,
  tag = "$BUILD_ID",
  region = cr_region_get(),
  bucket = cr_bucket_get(),
  projectId = cr_project_get(),
  launch_browser = interactive(),
  timeout = 600L
)
}
\arguments{
\item{html_folder}{the folder containing all the html}

\item{remote}{The folder on Google Cloud Storage, and the name of the service on Cloud Run}

\item{image_name}{The gcr.io image name that will be deployed and/or built}

\item{tag}{The tag to attached to the pushed image - can use \code{Build} macros}

\item{region}{The Cloud Run endpoint set by CR_REGION env arg}

\item{bucket}{The Cloud Storage bucket that will hold the code}

\item{projectId}{The projectId where it all gets deployed to}

\item{launch_browser}{Whether to launch the logs URL in a browser once deployed}

\item{timeout}{Amount of time that this build should be allowed to run, to second}
}
\description{
Supply the html folder to host it on Cloud Run.  Builds the dockerfile with the html within it, then deploys to Cloud Run
}
\details{
Will add a \code{default.template} file to the html folder that holds the nginx configuration
}
