% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{cr_deploy_pkgdown}
\alias{cr_deploy_pkgdown}
\title{Deploy a trigger for auto-builds a pkgdown website for an R package}
\usage{
cr_deploy_pkgdown(trigger, steps = NULL,
  git_email = "googlecloudrunner@r.com", keyring = "my-keyring",
  key = "github-key", env = NULL, substitutions = NULL,
  cipher = "id_rsa.enc",
  build_image = "gcr.io/gcer-public/packagetools:master")
}
\arguments{
\item{trigger}{The trigger source which will be a \link{RepoSource} or a \link{GitHubEventsConfig}}

\item{steps}{extra steps to run before the pkgdown website steps run}

\item{git_email}{The email the git commands will be identifying as}

\item{keyring}{The Key Management Store keyring containing the git ssh key}

\item{key}{The Key Management Store key containing the gitssh key}

\item{env}{A character vector of env arguments to set for all steps}

\item{substitutions}{A named list of Custom Build macros that can be substituted for values in the build steps.  Will be added to an existing default substitution \code{_$GIT_REPO} which holds the git repo as deployed in \code{trigger}}

\item{cipher}{The filename of the encrypted git ssh key that has been checked into the repository}

\item{build_image}{A docker image with \code{pkgdown} installed}
}
\description{
This will build a pkgdown website each time the trigger fires and deploy it to git
}
\details{
The trigger repository needs to hold an R package configured to build a pkgdown website.

For GitHub, the repository will also need to be linked to the project you are building within, via \url{https://console.cloud.google.com/cloud-build/triggers/connect}

The git ssh keys need to be deployed to Google KMS for the deployment of the website - see \link{cr_buildstep_git} - this only needs to be done once per Git account.  You then need to commit the encrypted ssh key (by default called \code{id_rsa.enc})
}
\examples{

\dontrun{

my_repo <- GitHubEventsConfig("MarkEdmondson1234/googleAnalyticsR")
cr_deploy_pkgdown(my_repo)

}
}
\seealso{
Create your own custom deployment using \link{cr_buildstep_pkgdown} which this function uses with some defaults
}
