% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{cr_dockerfile}
\alias{cr_dockerfile}
\title{Create Dockerfile in the deployment folder}
\usage{
cr_dockerfile(deploy_folder, ...)
}
\arguments{
\item{deploy_folder}{The folder containing the assessts to deploy}

\item{...}{Other arguments pass to \link[containerit]{dockerfile}}
}
\value{
An object of class Dockerfile
}
\description{
This uses \link[containerit]{dockerfile} to create a Dockerfile if possible
}
\examples{

\dontrun{
cr_dockerfile(system.file("example/", package = "googleCloudRunner"))
}
}
