% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{cr_email_set}
\alias{cr_email_set}
\title{Set cloud build email}
\usage{
cr_email_set(cloudbuildEmail)
}
\arguments{
\item{cloudbuildEmail}{The Cloud Build service email}
}
\description{
Needed so Cloud Scheduler can run Cloud Build jobs
}
\seealso{
https://console.cloud.google.com/cloud-build/settings
}
