% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudrun.R
\name{cr_endpoint_set}
\alias{cr_endpoint_set}
\title{Set the endpoint for your CloudRun services}
\usage{
cr_endpoint_set(region = c("us-central1", "asia-northeast1",
  "europe-west1", "us-east1"), override = NULL)
}
\arguments{
\item{region}{Region for the endpoint}

\item{override}{Set a full endpoint starting with https here if necessary}
}
\description{
Set the endpoint for your CloudRun services
}
