% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{cr_init}
\alias{cr_init}
\title{Set the endpoint for your CloudRun services}
\usage{
cr_init(region = c("europe-west1", "us-central1", "asia-northeast1",
  "us-east1"), override = NULL)
}
\arguments{
\item{region}{Region for the endpoint}

\item{override}{Set a full endpoint starting with https here if necessary}
}
\description{
Set the endpoint for your CloudRun services
}
