% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{cr_project_set}
\alias{cr_project_set}
\alias{cr_project_get}
\title{Get/Set the projectId for your CloudRun services}
\usage{
cr_project_set(projectId)

cr_project_get()
}
\arguments{
\item{projectId}{The projectId}
}
\description{
Can also use environment argument GCE_DEFAULT_PROJECT_ID
}
\examples{

cr_project_get()
}
