% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\name{cr_pubsub}
\alias{cr_pubsub}
\title{Send a message to pubsub}
\usage{
cr_pubsub(endpoint, payload = jsonlite::toJSON("hello"))
}
\arguments{
\item{endpoint}{The url endpoint of the PubSub service}

\item{payload}{Will be base64 encoded and placed in \code{message$data}}
}
\description{
Useful for testing Cloud Run pubsub deployments
}
