% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{cr_region_set}
\alias{cr_region_set}
\title{Set the endpoint for your CloudRun services}
\usage{
cr_region_set(region = c("europe-west1", "us-central1",
  "asia-northeast1", "us-east1"))
}
\arguments{
\item{region}{Region for the endpoint}
}
\description{
Can also use environment argument CR_REGION
}
