% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_regions.R
\docType{data}
\name{cr_regions}
\alias{cr_regions}
\title{Cloud Run Regions}
\format{
A character vector of valid Cloud Run region names
}
\usage{
cr_regions
}
\description{
Cloud Run Regions
}
\keyword{datasets}
