% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudrun_schedule.R
\name{cr_run_email}
\alias{cr_run_email}
\title{Create an invoker email for use within authenticated Cloud Run}
\usage{
cr_run_email(name, projectId = cr_project_get())
}
\arguments{
\item{name}{Name of the Cloud Run service}

\item{projectId}{The projectId where the Cloud Run service will run - set to NULL to only return the processed service name}
}
\description{
Create an invoker email for use within authenticated Cloud Run
}
\examples{

cr_run_email("my-run-app", "my-project")
}
\seealso{
Other Cloud Run functions: 
\code{\link{cr_jwt_create}()},
\code{\link{cr_plumber_pubsub}()},
\code{\link{cr_run_get}()},
\code{\link{cr_run_list}()},
\code{\link{cr_run_schedule_http}()},
\code{\link{cr_run}()}
}
\concept{Cloud Run functions}
