% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudrun.R
\name{cr_run_get}
\alias{cr_run_get}
\title{Get information about a Cloud Run service.}
\usage{
cr_run_get(name, projectId = Sys.getenv("GCE_DEFAULT_PROJECT_ID"))
}
\arguments{
\item{name}{The name of the service to retrieve}
}
\description{
Get information about a Cloud Run service.
}
\details{

}
\seealso{
\href{https://cloud.google.com/run/docs/reference/rest/v1/namespaces.services/get}{Google Documentation on namespaces.services.get}
}
