% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudrun.R
\name{cr_run_list}
\alias{cr_run_list}
\title{List CloudRun services.}
\usage{
cr_run_list(project, labelSelector = NULL, limit = NULL)
}
\arguments{
\item{labelSelector}{Allows to filter resources based on a label}

\item{limit}{The maximum number of records that should be returned}

\item{parent}{The GCP project from which the services should be listed}
}
\description{
List CloudRun services.
}
\details{

}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}
}
