% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudrun_schedule.R
\name{cr_run_schedule_http}
\alias{cr_run_schedule_http}
\title{Create a Cloud Scheduler HTTP target for a Cloud Run URI}
\usage{
cr_run_schedule_http(
  uri,
  http_method = "GET",
  body = NULL,
  email = cr_email_get()
)
}
\arguments{
\item{uri}{The URI of your Cloud Run application}

\item{http_method}{The HTTP verb you have set up your Cloud Run application to receive}

\item{body}{(optional) An R list object that will be turned into JSON via \link[jsonlite]{toJSON} and turned into a base64-encoded string if you are doing a POST, PUT or PATCH request.}

\item{email}{The email that will authenticate the job set via \link{cr_email_set}}
}
\value{
A \link{HttpTarget} object for use in \link{cr_schedule}
}
\description{
This enables Cloud Scheduler to trigger Cloud Run
}
\details{
Ensure you have a service email with \link{cr_email_set} of format \code{service-{project-number}@gcp-sa-cloudscheduler.iam.gserviceaccount.com} with Cloud Scheduler Service Agent role as per https://cloud.google.com/scheduler/docs/http-target-auth#add
}
\examples{

run_me <- cr_run_schedule_http(
  "https://example-ewjogewawq-ew.a.run.app/echo?msg=blah",
  http_method = "GET"
)

\dontrun{

cr_schedule("cloud-run-scheduled", schedule = "16 4 * * *",
            httpTarget = run_me)

}

}
\seealso{
https://cloud.google.com/run/docs/triggering/using-scheduler

\link{cr_build_schedule_http}

Other Cloud Scheduler functions: 
\code{\link{HttpTarget}()},
\code{\link{Job}()},
\code{\link{cr_build_schedule_http}()},
\code{\link{cr_schedule_delete}()},
\code{\link{cr_schedule_get}()},
\code{\link{cr_schedule_list}()},
\code{\link{cr_schedule_pause}()},
\code{\link{cr_schedule_run}()},
\code{\link{cr_schedule}()}
}
\concept{Cloud Scheduler functions}
