% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{cr_setup}
\alias{cr_setup}
\title{A helper setup function for setting up use with googleCloudRunner}
\usage{
cr_setup()
}
\description{
A helper setup function for setting up use with googleCloudRunner
}
\seealso{
Other setup functions: 
\code{\link{cr_setup_auth}()},
\code{\link{cr_setup_service}()},
\code{\link{cr_setup_test}()}
}
\concept{setup functions}
