% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{cr_setup_auth}
\alias{cr_setup_auth}
\title{Create a service account for googleCloudRunner}
\usage{
cr_setup_auth(
  email = Sys.getenv("GARGLE_EMAIL"),
  file = "googlecloudrunner-auth-key.json"
)
}
\arguments{
\item{email}{What email to open OAuth2 with}

\item{file}{Where to save the authentication file}

\item{accountId}{Name of the service account in GCP IAM}
}
\value{
TRUE if the file is ready to be setup by \link{cr_setup}
}
\description{
This will use your Google OAuth2 user to create a suitable service account
}
