% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_buildemail.R
\name{cr_setup_service}
\alias{cr_setup_service}
\title{Give a service account the right permissions for googleCloudRunner operations}
\usage{
cr_setup_service(
  account_email,
  json = Sys.getenv("GAR_CLIENT_JSON"),
  email = Sys.getenv("GARGLE_EMAIL")
)
}
\arguments{
\item{account_email}{The service account email e.g. \code{accountId@projectid.iam.gserviceaccount.com}}

\item{json}{the project clientId JSON}

\item{email}{the email of an Owner/Editor for the project}
}
\description{
Give a service account the right permissions for googleCloudRunner operations
}
