% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tests.R
\name{cr_setup_test}
\alias{cr_setup_test}
\title{Run tests over your setup}
\usage{
cr_setup_test()
}
\description{
This allows you to check if your setup works - run \link{cr_setup} first.
}
\seealso{
Other setup functions: 
\code{\link{cr_setup_auth}()},
\code{\link{cr_setup_service}()},
\code{\link{cr_setup}()}
}
\concept{setup functions}
