% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-repos.R
\name{cr_sourcerepo_list}
\alias{cr_sourcerepo_list}
\title{List source repositories available under a project}
\usage{
cr_sourcerepo_list(projectId = cr_project_get())
}
\arguments{
\item{projectId}{The projectId that holds the repositories}
}
\description{
List source repositories available under a project
}
