% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{DomainMapping}
\alias{DomainMapping}
\title{DomainMapping Object}
\usage{
DomainMapping(status = NULL, apiVersion = NULL, spec = NULL,
  metadata = NULL)
}
\arguments{
\item{status}{The current status of the DomainMapping}

\item{apiVersion}{The API version for this call such as 'domains}

\item{spec}{The spec for this DomainMapping}

\item{metadata}{Metadata associated with this BuildTemplate}
}
\value{
DomainMapping object
}
\description{
DomainMapping Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Resource to hold the state and status of a user's domain mapping.
}
\seealso{
Other DomainMapping functions: \code{\link{namespaces.domainmappings.create}},
  \code{\link{namespaces.domainmappings.replaceDomainMapping}},
  \code{\link{projects.locations.domainmappings.create}},
  \code{\link{projects.locations.domainmappings.replaceDomainMapping}}
}
\concept{DomainMapping functions}
