% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{DomainMappingSpec}
\alias{DomainMappingSpec}
\title{DomainMappingSpec Object}
\usage{
DomainMappingSpec(forceOverride = NULL, routeName = NULL,
  certificateMode = NULL)
}
\arguments{
\item{forceOverride}{If set, the mapping will override any mapping set before this spec was set}

\item{routeName}{The name of the Knative Route that this DomainMapping applies to}

\item{certificateMode}{The mode of the certificate}
}
\value{
DomainMappingSpec object
}
\description{
DomainMappingSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The desired state of the Domain Mapping.
}
\concept{DomainMappingSpec functions}
