% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{DomainMappingStatus}
\alias{DomainMappingStatus}
\title{DomainMappingStatus Object}
\usage{
DomainMappingStatus(mappedRouteName = NULL, conditions = NULL,
  observedGeneration = NULL, resourceRecords = NULL)
}
\arguments{
\item{mappedRouteName}{The name of the route that the mapping currently points to}

\item{conditions}{Array of observed DomainMappingConditions, indicating the current state}

\item{observedGeneration}{ObservedGeneration is the 'Generation' of the DomainMapping that}

\item{resourceRecords}{The resource records required to configure this domain mapping}
}
\value{
DomainMappingStatus object
}
\description{
DomainMappingStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The current state of the Domain Mapping.
}
\concept{DomainMappingStatus functions}
