% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{EnvFromSource}
\alias{EnvFromSource}
\title{EnvFromSource Object}
\usage{
EnvFromSource(configMapRef = NULL, secretRef = NULL, prefix = NULL)
}
\arguments{
\item{configMapRef}{(Optional)}

\item{secretRef}{(Optional)}

\item{prefix}{(Optional)}
}
\value{
EnvFromSource object
}
\description{
EnvFromSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedEnvFromSource represents the source of a set of ConfigMaps
}
\concept{EnvFromSource functions}
