% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{EnvVar}
\alias{EnvVar}
\title{EnvVar Object}
\usage{
EnvVar(value = NULL, name = NULL, valueFrom = NULL)
}
\arguments{
\item{value}{(Optional)}

\item{name}{Name of the environment variable}

\item{valueFrom}{(Optional)}
}
\value{
EnvVar object
}
\description{
EnvVar Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
EnvVar represents an environment variable present in a Container.
}
\concept{EnvVar functions}
