% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{EnvVarSource}
\alias{EnvVarSource}
\title{EnvVarSource Object}
\usage{
EnvVarSource(configMapKeyRef = NULL, secretKeyRef = NULL)
}
\arguments{
\item{configMapKeyRef}{(Optional)}

\item{secretKeyRef}{(Optional)}
}
\value{
EnvVarSource object
}
\description{
EnvVarSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedEnvVarSource represents a source for the value of an EnvVar.
}
\concept{EnvVarSource functions}
