% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{Expr}
\alias{Expr}
\title{Expr Object}
\usage{
Expr(expression = NULL, title = NULL, location = NULL,
  description = NULL)
}
\arguments{
\item{expression}{Textual representation of an expression in}

\item{title}{An optional title for the expression, i}

\item{location}{An optional string indicating the location of the expression for error}

\item{description}{An optional description of the expression}
}
\value{
Expr object
}
\description{
Expr Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an expression text. Example:    title: 'User account presence'    description: 'Determines whether the request has a user account'    expression: 'size(request.user) > 0'
}
\concept{Expr functions}
