% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{FileHashes}
\alias{FileHashes}
\title{FileHashes Object}
\usage{
FileHashes(fileHash = NULL)
}
\arguments{
\item{fileHash}{Collection of file hashes}
}
\value{
FileHashes object
}
\description{
FileHashes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Container message for hashes of byte content of files, used inSourceProvenance messages to verify integrity of source input to the build.
}
\concept{FileHashes functions}
