% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{GitHubEventsConfig}
\alias{GitHubEventsConfig}
\title{GitHubEventsConfig Object}
\usage{
GitHubEventsConfig(pullRequest = NULL, push = NULL,
  installationId = NULL, owner = NULL, name = NULL)
}
\arguments{
\item{pullRequest}{filter to match changes in pull requests}

\item{push}{filter to match changes in refs like branches, tags}

\item{installationId}{The installationID that emits the GitHub event}

\item{owner}{Owner of the repository}

\item{name}{Name of the repository}
}
\value{
GitHubEventsConfig object
}
\description{
GitHubEventsConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
GitHubEventsConfig describes the configuration of a trigger that creates abuild whenever a GitHub event is received.This message is experimental.
}
\concept{GitHubEventsConfig functions}
