% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtrigger_objects.R
\name{GitRepoSource}
\alias{GitRepoSource}
\title{GitRepoSource
Used for PubSub triggers}
\usage{
GitRepoSource(uri, ref, repoType = c("GITHUB", "CLOUD_SOURCE_REPOSITORIES"))
}
\description{
GitRepoSource
Used for PubSub triggers
}
