% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{GoogleCloudRunV1Condition}
\alias{GoogleCloudRunV1Condition}
\title{GoogleCloudRunV1Condition Object}
\usage{
GoogleCloudRunV1Condition(status = NULL, severity = NULL,
  reason = NULL, type = NULL, message = NULL,
  lastTransitionTime = NULL)
}
\arguments{
\item{status}{Status of the condition, one of True, False, Unknown}

\item{severity}{Optional}

\item{reason}{Optional}

\item{type}{type is used to communicate the status of the reconciliation process}

\item{message}{Optional}

\item{lastTransitionTime}{Optional}
}
\value{
GoogleCloudRunV1Condition object
}
\description{
GoogleCloudRunV1Condition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Condition defines a generic condition for a Resource
}
\concept{GoogleCloudRunV1Condition functions}
