% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{Handler}
\alias{Handler}
\title{Handler Object}
\usage{
Handler(tcpSocket = NULL, exec = NULL, httpGet = NULL)
}
\arguments{
\item{tcpSocket}{(Optional)}

\item{exec}{(Optional)}

\item{httpGet}{(Optional)}
}
\value{
Handler object
}
\description{
Handler Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedHandler defines a specific action that should be taken
}
\concept{Handler functions}
