% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{Hash}
\alias{Hash}
\title{Hash Object}
\usage{
Hash(type = NULL, value = NULL)
}
\arguments{
\item{type}{The type of hash that was performed}

\item{value}{The hash value}
}
\value{
Hash object
}
\description{
Hash Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Container message for hash values.
}
\concept{Hash functions}
