% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{HTTPGetAction}
\alias{HTTPGetAction}
\title{HTTPGetAction Object}
\usage{
HTTPGetAction(path = NULL, host = NULL, scheme = NULL,
  httpHeaders = NULL)
}
\arguments{
\item{path}{(Optional)}

\item{host}{(Optional)}

\item{scheme}{(Optional)}

\item{httpHeaders}{(Optional)}
}
\value{
HTTPGetAction object
}
\description{
HTTPGetAction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedHTTPGetAction describes an action based on HTTP Get requests.
}
\concept{HTTPGetAction functions}
