% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{HTTPHeader}
\alias{HTTPHeader}
\title{HTTPHeader Object}
\usage{
HTTPHeader(name = NULL, value = NULL)
}
\arguments{
\item{name}{Cloud Run fully managed: not supported}

\item{value}{Cloud Run fully managed: not supported}
}
\value{
HTTPHeader object
}
\description{
HTTPHeader Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedHTTPHeader describes a custom header to be used in HTTP probes
}
\concept{HTTPHeader functions}
