% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{HttpTarget.headers}
\alias{HttpTarget.headers}
\title{HttpTarget.headers Object}
\usage{
HttpTarget.headers()
}
\value{
HttpTarget.headers object
}
\description{
HttpTarget.headers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The user can specify HTTP request headers to send with the job'sHTTP request. This map contains the header field names andvalues. Repeated headers are not supported, but a header value cancontain commas. These headers represent a subset of the headersthat will accompany the job's HTTP request. Some HTTP requestheaders will be ignored or replaced. A partial list of headers thatwill be ignored or replaced is below:- Host: This will be computed by Cloud Scheduler and derived fromuri.* `Content-Length`: This will be computed by Cloud Scheduler.* `User-Agent`: This will be set to `'Google-Cloud-Scheduler'`.* `X-Google-*`: Google internal use only.* `X-AppEngine-*`: Google internal use only.The total size of headers must be less than 80KB.
}
\seealso{
Other HttpTarget functions: \code{\link{HttpTarget}}
}
\concept{HttpTarget functions}
