% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{HttpTarget}
\alias{HttpTarget}
\title{HttpTarget Object}
\usage{
HttpTarget(HttpTarget.headers = NULL, body = NULL, oauthToken = NULL,
  uri = NULL, headers = NULL, oidcToken = NULL, httpMethod = NULL)
}
\arguments{
\item{HttpTarget.headers}{The \link{HttpTarget.headers} object or list of objects}

\item{body}{HTTP request body}

\item{oauthToken}{If specified, an}

\item{uri}{Required}

\item{headers}{The user can specify HTTP request headers to send with the job's}

\item{oidcToken}{If specified, an}

\item{httpMethod}{Which HTTP method to use for the request}
}
\value{
HttpTarget object
}
\description{
HttpTarget Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Http target. The job will be pushed to the job handler by means ofan HTTP request via an http_method such as HTTPPOST, HTTP GET, etc. The job is acknowledged by means of an HTTPresponse code in the range [200 - 299]. A failure to receive a responseconstitutes a failed execution. For a redirected request, the responsereturned by the redirected request is considered.
}
\seealso{
Other HttpTarget functions: \code{\link{HttpTarget.headers}}
}
\concept{HttpTarget functions}
