% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{IntOrString}
\alias{IntOrString}
\title{IntOrString Object}
\usage{
IntOrString(type = NULL, intVal = NULL, strVal = NULL)
}
\arguments{
\item{type}{The type of the value}

\item{intVal}{The int value}

\item{strVal}{The string value}
}
\value{
IntOrString object
}
\description{
IntOrString Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
IntOrString is a type that can hold an int32 or a string.  When used inJSON or YAML marshalling and unmarshalling, it produces or consumes theinner type.  This allows you to have, for example, a JSON field that canaccept a name or number.
}
\concept{IntOrString functions}
