% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ListAuthorizedDomainsResponse}
\alias{ListAuthorizedDomainsResponse}
\title{ListAuthorizedDomainsResponse Object}
\usage{
ListAuthorizedDomainsResponse(nextPageToken = NULL, domains = NULL)
}
\arguments{
\item{nextPageToken}{Continuation token for fetching the next page of results}

\item{domains}{The authorized domains belonging to the user}
}
\value{
ListAuthorizedDomainsResponse object
}
\description{
ListAuthorizedDomainsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of Authorized Domains.
}
\concept{ListAuthorizedDomainsResponse functions}
