% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ListAutoDomainMappingsResponse}
\alias{ListAutoDomainMappingsResponse}
\title{ListAutoDomainMappingsResponse Object}
\usage{
ListAutoDomainMappingsResponse(unreachable = NULL, metadata = NULL,
  items = NULL, apiVersion = NULL)
}
\arguments{
\item{unreachable}{Locations that could not be reached}

\item{metadata}{Metadata associated with this AutoDomainMapping list}

\item{items}{List of AutoDomainMappings}

\item{apiVersion}{The API version for this call such as 'serving}
}
\value{
ListAutoDomainMappingsResponse object
}
\description{
ListAutoDomainMappingsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ListAutoDomainMappingsResponse is a list of AutoDomainMapping resources.
}
\concept{ListAutoDomainMappingsResponse functions}
