% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{ListBuildsResponse}
\alias{ListBuildsResponse}
\title{ListBuildsResponse Object}
\usage{
ListBuildsResponse(nextPageToken = NULL, builds = NULL)
}
\arguments{
\item{nextPageToken}{Token to receive the next page of results}

\item{builds}{Builds will be sorted by `create_time`, descending}
}
\value{
ListBuildsResponse object
}
\description{
ListBuildsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response including listed builds.
}
\concept{ListBuildsResponse functions}
