% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ListConfigurationsResponse}
\alias{ListConfigurationsResponse}
\title{ListConfigurationsResponse Object}
\usage{
ListConfigurationsResponse(unreachable = NULL, metadata = NULL,
  items = NULL, apiVersion = NULL)
}
\arguments{
\item{unreachable}{Locations that could not be reached}

\item{metadata}{Metadata associated with this Configuration list}

\item{items}{List of Configurations}

\item{apiVersion}{The API version for this call such as 'serving}
}
\value{
ListConfigurationsResponse object
}
\description{
ListConfigurationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ListConfigurationsResponse is a list of Configuration resources.
}
\concept{ListConfigurationsResponse functions}
