% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ListDomainMappingsResponse}
\alias{ListDomainMappingsResponse}
\title{ListDomainMappingsResponse Object}
\usage{
ListDomainMappingsResponse(unreachable = NULL, metadata = NULL,
  items = NULL, apiVersion = NULL)
}
\arguments{
\item{unreachable}{Locations that could not be reached}

\item{metadata}{Metadata associated with this DomainMapping list}

\item{items}{List of DomainMappings}

\item{apiVersion}{The API version for this call such as 'domains}
}
\value{
ListDomainMappingsResponse object
}
\description{
ListDomainMappingsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ListDomainMappingsResponse is a list of DomainMapping resources.
}
\concept{ListDomainMappingsResponse functions}
