% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ListMeta}
\alias{ListMeta}
\title{ListMeta Object}
\usage{
ListMeta(continue = NULL, selfLink = NULL, resourceVersion = NULL)
}
\arguments{
\item{continue}{continue may be set if the user set a limit on the number of items}

\item{selfLink}{SelfLink is a URL representing this object}

\item{resourceVersion}{String that identifies the server's internal version of this object that}
}
\value{
ListMeta object
}
\description{
ListMeta Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ListMeta describes metadata that synthetic resources must have, includinglists and various status objects. A resource may have only one of{ObjectMeta, ListMeta}.
}
\concept{ListMeta functions}
