% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{ListOperationsResponse}
\alias{ListOperationsResponse}
\title{ListOperationsResponse Object}
\usage{
ListOperationsResponse(operations = NULL, nextPageToken = NULL)
}
\arguments{
\item{operations}{A list of operations that matches the specified filter in the request}

\item{nextPageToken}{The standard List next-page token}
}
\value{
ListOperationsResponse object
}
\description{
ListOperationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for Operations.ListOperations.
}
\concept{ListOperationsResponse functions}
