% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ListRevisionsResponse}
\alias{ListRevisionsResponse}
\title{ListRevisionsResponse Object}
\usage{
ListRevisionsResponse(unreachable = NULL, metadata = NULL,
  items = NULL, apiVersion = NULL)
}
\arguments{
\item{unreachable}{Locations that could not be reached}

\item{metadata}{Metadata associated with this revision list}

\item{items}{List of Revisions}

\item{apiVersion}{The API version for this call such as 'serving}
}
\value{
ListRevisionsResponse object
}
\description{
ListRevisionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ListRevisionsResponse is a list of Revision resources.
}
\concept{ListRevisionsResponse functions}
