% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ListRoutesResponse}
\alias{ListRoutesResponse}
\title{ListRoutesResponse Object}
\usage{
ListRoutesResponse(unreachable = NULL, metadata = NULL, items = NULL,
  apiVersion = NULL)
}
\arguments{
\item{unreachable}{Locations that could not be reached}

\item{metadata}{Metadata associated with this Route list}

\item{items}{List of Routes}

\item{apiVersion}{The API version for this call such as 'serving}
}
\value{
ListRoutesResponse object
}
\description{
ListRoutesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ListRoutesResponse is a list of Route resources.
}
\concept{ListRoutesResponse functions}
