% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ListServicesResponse}
\alias{ListServicesResponse}
\title{ListServicesResponse Object}
\usage{
ListServicesResponse(apiVersion = NULL, unreachable = NULL,
  metadata = NULL, items = NULL)
}
\arguments{
\item{apiVersion}{The API version for this call such as 'serving}

\item{unreachable}{Locations that could not be reached}

\item{metadata}{Metadata associated with this Service list}

\item{items}{List of Services}
}
\value{
ListServicesResponse object
}
\description{
ListServicesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of Service resources.
}
\concept{ListServicesResponse functions}
