% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{LocalObjectReference}
\alias{LocalObjectReference}
\title{LocalObjectReference Object}
\usage{
LocalObjectReference(name = NULL)
}
\arguments{
\item{name}{(Optional)}
}
\value{
LocalObjectReference object
}
\description{
LocalObjectReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedLocalObjectReference contains enough information to let you locate thereferenced object inside the same namespace.
}
\concept{LocalObjectReference functions}
