% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Location.labels}
\alias{Location.labels}
\title{Location.labels Object}
\usage{
Location.labels()
}
\value{
Location.labels object
}
\description{
Location.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cross-service attributes for the location. For example    {'cloud.googleapis.com/region': 'us-east1'}
}
\seealso{
Other Location functions: \code{\link{Location.metadata}},
  \code{\link{Location}}
}
\concept{Location functions}
