% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{namespaces.authorizeddomains.list}
\alias{namespaces.authorizeddomains.list}
\title{List authorized domains.}
\usage{
namespaces.authorizeddomains.list(parent, pageToken = NULL,
  pageSize = NULL)
}
\arguments{
\item{parent}{Name of the parent Application resource}

\item{pageToken}{Continuation token for fetching the next page of results}

\item{pageSize}{Maximum results to return per page}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}
}
