% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{namespaces.autodomainmappings.create}
\alias{namespaces.autodomainmappings.create}
\title{Creates a new auto domain mapping.}
\usage{
namespaces.autodomainmappings.create(AutoDomainMapping, parent)
}
\arguments{
\item{AutoDomainMapping}{The \link{AutoDomainMapping} object to pass to this method}

\item{parent}{The project ID or project number in which this auto domain mapping should}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}

Other AutoDomainMapping functions: \code{\link{AutoDomainMapping}},
  \code{\link{namespaces.autodomainmappings.replaceAutoDomainMapping}},
  \code{\link{projects.locations.autodomainmappings.create}},
  \code{\link{projects.locations.autodomainmappings.replaceAutoDomainMapping}}
}
\concept{AutoDomainMapping functions}
