% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{namespaces.autodomainmappings.delete}
\alias{namespaces.autodomainmappings.delete}
\title{Delete an auto domain mapping.}
\usage{
namespaces.autodomainmappings.delete(name, propagationPolicy = NULL,
  kind = NULL, apiVersion = NULL)
}
\arguments{
\item{name}{The name of the auto domain mapping being deleted}

\item{propagationPolicy}{Specifies the propagation policy of delete}

\item{kind}{Cloud Run currently ignores this parameter}

\item{apiVersion}{Cloud Run currently ignores this parameter}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}
}
