% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{namespaces.autodomainmappings.replaceAutoDomainMapping}
\alias{namespaces.autodomainmappings.replaceAutoDomainMapping}
\title{Replace an auto domain mapping.Only the spec and metadata labels and annotations are modifiable. Afterthe Update request, Cloud Run will work to make the 'status'match the requested 'spec'.May provide metadata.resourceVersion to enforce update from last read foroptimistic concurrency control.}
\usage{
namespaces.autodomainmappings.replaceAutoDomainMapping(AutoDomainMapping,
  name)
}
\arguments{
\item{AutoDomainMapping}{The \link{AutoDomainMapping} object to pass to this method}

\item{name}{The name of the auto domain mapping being retrieved}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}

Other AutoDomainMapping functions: \code{\link{namespaces.autodomainmappings.create}},
  \code{\link{projects.locations.autodomainmappings.create}},
  \code{\link{projects.locations.autodomainmappings.replaceAutoDomainMapping}}
}
\concept{AutoDomainMapping functions}
