% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{namespaces.configurations.list}
\alias{namespaces.configurations.list}
\title{List configurations.}
\usage{
namespaces.configurations.list(parent, continue = NULL, limit = NULL,
  watch = NULL, labelSelector = NULL, resourceVersion = NULL,
  includeUninitialized = NULL, fieldSelector = NULL)
}
\arguments{
\item{parent}{The namespace from which the configurations should be listed}

\item{continue}{Optional encoded string to continue paging}

\item{limit}{The maximum number of records that should be returned}

\item{watch}{Flag that indicates that the client expects to watch this resource as well}

\item{labelSelector}{Allows to filter resources based on a label}

\item{resourceVersion}{The baseline resource version from which the list or watch operation should}

\item{includeUninitialized}{Not currently used by Cloud Run}

\item{fieldSelector}{Allows to filter resources based on a specific value for a field name}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}
}
