% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{namespaces.revisions.list}
\alias{namespaces.revisions.list}
\title{List revisions.}
\usage{
namespaces.revisions.list(parent, labelSelector = NULL,
  resourceVersion = NULL, includeUninitialized = NULL,
  fieldSelector = NULL, continue = NULL, limit = NULL,
  watch = NULL)
}
\arguments{
\item{parent}{The namespace from which the revisions should be listed}

\item{labelSelector}{Allows to filter resources based on a label}

\item{resourceVersion}{The baseline resource version from which the list or watch operation should}

\item{includeUninitialized}{Not currently used by Cloud Run}

\item{fieldSelector}{Allows to filter resources based on a specific value for a field name}

\item{continue}{Optional encoded string to continue paging}

\item{limit}{The maximum number of records that should be returned}

\item{watch}{Flag that indicates that the client expects to watch this resource as well}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}
}
