% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{namespaces.routes.create}
\alias{namespaces.routes.create}
\title{Create a route.}
\usage{
namespaces.routes.create(Route, parent)
}
\arguments{
\item{Route}{The \link{Route} object to pass to this method}

\item{parent}{The namespace in which the route should be created}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}

Other Route functions: \code{\link{Route}},
  \code{\link{namespaces.routes.replaceRoute}},
  \code{\link{projects.locations.routes.create}},
  \code{\link{projects.locations.routes.replaceRoute}}
}
\concept{Route functions}
