% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{namespaces.services.replaceService}
\alias{namespaces.services.replaceService}
\title{Replace a service.Only the spec and metadata labels and annotations are modifiable. Afterthe Update request, Cloud Run will work to make the 'status'match the requested 'spec'.May provide metadata.resourceVersion to enforce update from last read foroptimistic concurrency control.}
\usage{
namespaces.services.replaceService(Service, name)
}
\arguments{
\item{Service}{The \link{Service} object to pass to this method}

\item{name}{The name of the service being replaced}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}

Other Service functions: \code{\link{Service}},
  \code{\link{namespaces.services.create}},
  \code{\link{projects.locations.services.create}},
  \code{\link{projects.locations.services.replaceService}}
}
\concept{Service functions}
