% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{OAuthToken}
\alias{OAuthToken}
\title{OAuthToken Object}
\usage{
OAuthToken(scope = NULL, serviceAccountEmail = NULL)
}
\arguments{
\item{scope}{OAuth scope to be used for generating OAuth access token}

\item{serviceAccountEmail}{[Service account email](https://cloud}
}
\value{
OAuthToken object
}
\description{
OAuthToken Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information needed for generating an[OAuth token](https://developers.google.com/identity/protocols/OAuth2).This type of authorization should generally only be used when calling GoogleAPIs hosted on *.googleapis.com.
}
\concept{OAuthToken functions}
