% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ObjectMeta.annotations}
\alias{ObjectMeta.annotations}
\title{ObjectMeta.annotations Object}
\usage{
ObjectMeta.annotations()
}
\value{
ObjectMeta.annotations object
}
\description{
ObjectMeta.annotations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
(Optional)Annotations is an unstructured key value map stored with a resource thatmay be set by external tools to store and retrieve arbitrary metadata. Theyare not queryable and should be preserved when modifying objects. Moreinfo: http://kubernetes.io/docs/user-guide/annotations
}
\seealso{
Other ObjectMeta functions: \code{\link{ObjectMeta.labels}},
  \code{\link{ObjectMeta}}
}
\concept{ObjectMeta functions}
