% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ObjectMeta.labels}
\alias{ObjectMeta.labels}
\title{ObjectMeta.labels Object}
\usage{
ObjectMeta.labels()
}
\value{
ObjectMeta.labels object
}
\description{
ObjectMeta.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
(Optional)Map of string keys and values that can be used to organize and categorize(scope and select) objects. May match selectors of replication controllersand routes.More info: http://kubernetes.io/docs/user-guide/labels
}
\seealso{
Other ObjectMeta functions: \code{\link{ObjectMeta.annotations}},
  \code{\link{ObjectMeta}}
}
\concept{ObjectMeta functions}
