% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ObjectMeta}
\alias{ObjectMeta}
\title{ObjectMeta Object}
\usage{
ObjectMeta(ObjectMeta.labels = NULL, ObjectMeta.annotations = NULL,
  creationTimestamp = NULL, labels = NULL, generation = NULL,
  resourceVersion = NULL, selfLink = NULL, uid = NULL,
  namespace = NULL, annotations = NULL, generateName = NULL,
  clusterName = NULL, deletionGracePeriodSeconds = NULL, name = NULL,
  finalizers = NULL, deletionTimestamp = NULL,
  ownerReferences = NULL)
}
\arguments{
\item{ObjectMeta.labels}{The \link{ObjectMeta.labels} object or list of objects}

\item{ObjectMeta.annotations}{The \link{ObjectMeta.annotations} object or list of objects}

\item{creationTimestamp}{(Optional)}

\item{labels}{(Optional)}

\item{generation}{(Optional)}

\item{resourceVersion}{(Optional)}

\item{selfLink}{(Optional)}

\item{uid}{(Optional)}

\item{namespace}{Namespace defines the space within each name must be unique, within a}

\item{annotations}{(Optional)}

\item{generateName}{(Optional)}

\item{clusterName}{(Optional)}

\item{deletionGracePeriodSeconds}{(Optional)}

\item{name}{Name must be unique within a namespace, within a Cloud Run region}

\item{finalizers}{(Optional)}

\item{deletionTimestamp}{(Optional)}

\item{ownerReferences}{(Optional)}
}
\value{
ObjectMeta object
}
\description{
ObjectMeta Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
k8s.io.apimachinery.pkg.apis.meta.v1.ObjectMeta is metadata that allpersisted resources must have, which includes all objects users must create.
}
\seealso{
Other ObjectMeta functions: \code{\link{ObjectMeta.annotations}},
  \code{\link{ObjectMeta.labels}}
}
\concept{ObjectMeta functions}
