% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{OidcToken}
\alias{OidcToken}
\title{OidcToken Object}
\usage{
OidcToken(audience = NULL, serviceAccountEmail = NULL)
}
\arguments{
\item{audience}{Audience to be used when generating OIDC token}

\item{serviceAccountEmail}{[Service account email](https://cloud}
}
\value{
OidcToken object
}
\description{
OidcToken Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information needed for generating an[OpenID Connecttoken](https://developers.google.com/identity/protocols/OpenIDConnect).This type of authorization can be used for many scenarios, includingcalling Cloud Run, or endpoints where you intend to validate the tokenyourself.
}
\concept{OidcToken functions}
