% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{Operation.metadata}
\alias{Operation.metadata}
\title{Operation.metadata Object}
\usage{
Operation.metadata()
}
\value{
Operation.metadata object
}
\description{
Operation.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Service-specific metadata associated with the operation.  It typicallycontains progress information and common metadata such as create time.Some services might not provide such metadata.  Any method that returns along-running operation should document the metadata type, if any.
}
\seealso{
Other Operation functions: \code{\link{Operation.response}},
  \code{\link{Operation}}
}
\concept{Operation functions}
