% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{Operation.response}
\alias{Operation.response}
\title{Operation.response Object}
\usage{
Operation.response()
}
\value{
Operation.response object
}
\description{
Operation.response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The normal response of the operation in case of success.  If the originalmethod returns no data on success, such as `Delete`, the response is`google.protobuf.Empty`.  If the original method is standard`Get`/`Create`/`Update`, the response should be the resource.  For othermethods, the response should have the type `XxxResponse`, where `Xxx`is the original method name.  For example, if the original method nameis `TakeSnapshot()`, the inferred response type is`TakeSnapshotResponse`.
}
\seealso{
Other Operation functions: \code{\link{Operation.metadata}},
  \code{\link{Operation}}
}
\concept{Operation functions}
