% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{OwnerReference}
\alias{OwnerReference}
\title{OwnerReference Object}
\usage{
OwnerReference(uid = NULL, controller = NULL,
  blockOwnerDeletion = NULL, apiVersion = NULL, name = NULL)
}
\arguments{
\item{uid}{UID of the referent}

\item{controller}{If true, this reference points to the managing controller}

\item{blockOwnerDeletion}{If true, AND if the owner has the 'foregroundDeletion' finalizer, then}

\item{apiVersion}{API version of the referent}

\item{name}{Name of the referent}
}
\value{
OwnerReference object
}
\description{
OwnerReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
OwnerReference contains enough information to let you identify an owningobject. Currently, an owning object must be in the same namespace, so thereis no namespace field.
}
\concept{OwnerReference functions}
