% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{Policy}
\alias{Policy}
\title{Policy Object}
\usage{
Policy(etag = NULL, version = NULL, auditConfigs = NULL,
  bindings = NULL)
}
\arguments{
\item{etag}{`etag` is used for optimistic concurrency control as a way to help}

\item{version}{Specifies the format of the policy}

\item{auditConfigs}{Specifies cloud audit logging configuration for this policy}

\item{bindings}{Associates a list of `members` to a `role`}
}
\value{
Policy object
}
\description{
Policy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines an Identity and Access Management (IAM) policy. It is used tospecify access control policies for Cloud Platform resources.A `Policy` is a collection of `bindings`. A `binding` binds one or more`members` to a single `role`. Members can be user accounts, service accounts,Google groups, and domains (such as G Suite). A `role` is a named list ofpermissions (defined by IAM or configured by users). A `binding` canoptionally specify a `condition`, which is a logic expression that furtherconstrains the role binding based on attributes about the request and/ortarget resource.**JSON Example**    {      'bindings': [        {          'role': 'roles/resourcemanager.organizationAdmin',          'members': [            'user:mike@example.com',            'group:admins@example.com',            'domain:google.com',            'serviceAccount:my-project-id@appspot.gserviceaccount.com'          ]        },        {          'role': 'roles/resourcemanager.organizationViewer',          'members': ['user:eve@example.com'],          'condition': {            'title': 'expirable access',            'description': 'Does not grant access after Sep 2020',            'expression': 'request.time <            timestamp('2020-10-01T00:00:00.000Z')',          }        }      ]    }**YAML Example**    bindings:    - members:      - user:mike@example.com      - group:admins@example.com      - domain:google.com      - serviceAccount:my-project-id@appspot.gserviceaccount.com      role: roles/resourcemanager.organizationAdmin    - members:      - user:eve@example.com      role: roles/resourcemanager.organizationViewer      condition:        title: expirable access        description: Does not grant access after Sep 2020        expression: request.time < timestamp('2020-10-01T00:00:00.000Z')For a description of IAM and its features, see the[IAM developer's guide](https://cloud.google.com/iam/docs).
}
\concept{Policy functions}
