% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Probe}
\alias{Probe}
\title{Cloud Run API Objects 
Deploy and manage user provided container images that scale automatically based on HTTP traffic.}
\usage{
Probe(periodSeconds = NULL, failureThreshold = NULL,
  initialDelaySeconds = NULL, timeoutSeconds = NULL, handler = NULL,
  successThreshold = NULL)
}
\arguments{
\item{periodSeconds}{(Optional)}

\item{failureThreshold}{(Optional)}

\item{initialDelaySeconds}{(Optional)}

\item{timeoutSeconds}{(Optional)}

\item{handler}{Cloud Run fully managed: not supported}

\item{successThreshold}{(Optional)}
}
\value{
Probe object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2019-11-03 11:18:30
filename: R/objects.R
api_json: cr
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Probe Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedProbe describes a health check to be performed against a container todetermine whether it is alive or ready to receive traffic.
}
\concept{Probe functions}
