% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_build.R
\name{projects.builds.cancel}
\alias{projects.builds.cancel}
\title{Cancels a build in progress.}
\usage{
projects.builds.cancel(CancelBuildRequest, projectId, id)
}
\arguments{
\item{CancelBuildRequest}{The \link{CancelBuildRequest} object to pass to this method}

\item{projectId}{ID of the project}

\item{id}{ID of the build}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/cloud-build/docs/}{Google Documentation}
}
\concept{CancelBuildRequest functions}
