% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_build.R
\name{projects.builds.create}
\alias{projects.builds.create}
\title{Starts a build with the specified configuration.This method returns a long-running `Operation`, which includes the buildID. Pass the build ID to `GetBuild` to determine the build status (such as`SUCCESS` or `FAILURE`).}
\usage{
projects.builds.create(Build, projectId)
}
\arguments{
\item{Build}{The \link{Build} object to pass to this method}

\item{projectId}{ID of the project}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/cloud-build/docs/}{Google Documentation}

Other Build functions: \code{\link{Build.substitutions}},
  \code{\link{Build.timing}}, \code{\link{Build}},
  \code{\link{cr_build}}
}
\concept{Build functions}
