% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_build.R
\name{projects.builds.list}
\alias{projects.builds.list}
\title{Lists previously requested builds.Previously requested builds may still be in-progress, or may have finishedsuccessfully or unsuccessfully.}
\usage{
projects.builds.list(projectId, pageToken = NULL, pageSize = NULL,
  filter = NULL)
}
\arguments{
\item{projectId}{ID of the project}

\item{pageToken}{Token to provide to skip to a particular spot in the list}

\item{pageSize}{Number of results to return in the list}

\item{filter}{The raw filter text to constrain the results}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/cloud-build/docs/}{Google Documentation}
}
