% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_build.R
\name{projects.builds.retry}
\alias{projects.builds.retry}
\title{Creates a new build based on the specified build.This method creates a new build using the original build request, which mayor may not result in an identical build.For triggered builds:* Triggered builds resolve to a precise revision; therefore a retry of atriggered build will result in a build that uses the same revision.For non-triggered builds that specify `RepoSource`:* If the original build built from the tip of a branch, the retried buildwill build from the tip of that branch, which may not be the same revisionas the original build.* If the original build specified a commit sha or revision ID, the retriedbuild will use the identical source.For builds that specify `StorageSource`:* If the original build pulled source from Google Cloud Storage withoutspecifying the generation of the object, the new build will use the currentobject, which may be different from the original build source.* If the original build pulled source from Cloud Storage and specified thegeneration of the object, the new build will attempt to use the sameobject, which may or may not be available depending on the bucket'slifecycle management settings.}
\usage{
projects.builds.retry(RetryBuildRequest, projectId, id)
}
\arguments{
\item{RetryBuildRequest}{The \link{RetryBuildRequest} object to pass to this method}

\item{projectId}{ID of the project}

\item{id}{Build ID of the original build}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/cloud-build/docs/}{Google Documentation}
}
\concept{RetryBuildRequest functions}
