% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{projects.locations.configurations.create}
\alias{projects.locations.configurations.create}
\title{Create a configuration.}
\usage{
projects.locations.configurations.create(Configuration, parent)
}
\arguments{
\item{Configuration}{The \link{Configuration} object to pass to this method}

\item{parent}{The namespace in which the configuration should be created}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}

Other Configuration functions: \code{\link{Configuration}},
  \code{\link{namespaces.configurations.create}},
  \code{\link{namespaces.configurations.replaceConfiguration}},
  \code{\link{projects.locations.configurations.replaceConfiguration}}
}
\concept{Configuration functions}
