% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{projects.locations.domainmappings.create}
\alias{projects.locations.domainmappings.create}
\title{Create a new domain mapping.}
\usage{
projects.locations.domainmappings.create(DomainMapping, parent)
}
\arguments{
\item{DomainMapping}{The \link{DomainMapping} object to pass to this method}

\item{parent}{The namespace in which the domain mapping should be created}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}

Other DomainMapping functions: \code{\link{DomainMapping}},
  \code{\link{namespaces.domainmappings.create}},
  \code{\link{namespaces.domainmappings.replaceDomainMapping}},
  \code{\link{projects.locations.domainmappings.replaceDomainMapping}}
}
\concept{DomainMapping functions}
