% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R, R/funcs_schedule.R
\name{projects.locations.get}
\alias{projects.locations.get}
\title{Gets information about a location.}
\usage{
projects.locations.get(name)

projects.locations.get(name)
}
\arguments{
\item{name}{Resource name for the location}

\item{name}{Resource name for the location}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}

Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.

Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}

\href{https://cloud.google.com/scheduler/}{Google Documentation}
}
