% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_schedule.R
\name{projects.locations.jobs.create}
\alias{projects.locations.jobs.create}
\title{Creates a job.}
\usage{
projects.locations.jobs.create(Job, parent)
}
\arguments{
\item{Job}{The \link{Job} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/scheduler/}{Google Documentation}

Other Job functions: \code{\link{Job}},
  \code{\link{cr_schedule}},
  \code{\link{projects.locations.jobs.patch}}
}
\concept{Job functions}
