% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_schedule.R
\name{projects.locations.jobs.list}
\alias{projects.locations.jobs.list}
\title{Lists jobs.}
\usage{
projects.locations.jobs.list(parent, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{parent}{Required}

\item{pageToken}{A token identifying a page of results the server will return}

\item{pageSize}{Requested page size}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/scheduler/}{Google Documentation}
}
