% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_schedule.R
\name{projects.locations.jobs.patch}
\alias{projects.locations.jobs.patch}
\title{Updates a job.If successful, the updated Job is returned. If the job doesnot exist, `NOT_FOUND` is returned.If UpdateJob does not successfully return, it is possible for thejob to be in an Job.State.UPDATE_FAILED state. A job in this state maynot be executed. If this happens, retry the UpdateJob requestuntil a successful response is received.}
\usage{
projects.locations.jobs.patch(Job, name, updateMask = NULL)
}
\arguments{
\item{Job}{The \link{Job} object to pass to this method}

\item{name}{Optionally caller-specified in CreateJob, after}

\item{updateMask}{A  mask used to specify which fields of the job are being updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/scheduler/}{Google Documentation}

Other Job functions: \code{\link{Job}},
  \code{\link{cr_schedule}},
  \code{\link{projects.locations.jobs.create}}
}
\concept{Job functions}
