% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_schedule.R
\name{projects.locations.jobs.pause}
\alias{projects.locations.jobs.pause}
\title{Pauses a job.If a job is paused then the system will stop executing the jobuntil it is re-enabled via ResumeJob. Thestate of the job is stored in state; if paused itwill be set to Job.State.PAUSED. A job must be in Job.State.ENABLEDto be paused.}
\usage{
projects.locations.jobs.pause(PauseJobRequest, name)
}
\arguments{
\item{PauseJobRequest}{The \link{PauseJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/scheduler/}{Google Documentation}
}
\concept{PauseJobRequest functions}
