% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_schedule.R
\name{projects.locations.jobs.resume}
\alias{projects.locations.jobs.resume}
\title{Resume a job.This method reenables a job after it has been Job.State.PAUSED. Thestate of a job is stored in Job.state; after calling this method itwill be set to Job.State.ENABLED. A job must be inJob.State.PAUSED to be resumed.}
\usage{
projects.locations.jobs.resume(ResumeJobRequest, name)
}
\arguments{
\item{ResumeJobRequest}{The \link{ResumeJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/scheduler/}{Google Documentation}

Other ResumeJobRequest functions: \code{\link{ResumeJobRequest}}
}
\concept{ResumeJobRequest functions}
