% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_schedule.R
\name{projects.locations.jobs.run}
\alias{projects.locations.jobs.run}
\title{Forces a job to run now.When this method is called, Cloud Scheduler will dispatch the job, evenif the job is already running.}
\usage{
projects.locations.jobs.run(RunJobRequest, name)
}
\arguments{
\item{RunJobRequest}{The \link{RunJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/scheduler/}{Google Documentation}

Other RunJobRequest functions: \code{\link{RunJobRequest}}
}
\concept{RunJobRequest functions}
