% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{projects.locations.routes.delete}
\alias{projects.locations.routes.delete}
\title{Delete a route.This will cause the route to stop serving traffic.}
\usage{
projects.locations.routes.delete(name, apiVersion = NULL,
  propagationPolicy = NULL, kind = NULL)
}
\arguments{
\item{name}{The name of the route to delete}

\item{apiVersion}{Cloud Run currently ignores this parameter}

\item{propagationPolicy}{Specifies the propagation policy of delete}

\item{kind}{Cloud Run currently ignores this parameter}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}
}
