% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{projects.locations.routes.get}
\alias{projects.locations.routes.get}
\title{Get information about a route.}
\usage{
projects.locations.routes.get(name)
}
\arguments{
\item{name}{The name of the route to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}
}
