% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{projects.locations.services.create}
\alias{projects.locations.services.create}
\title{Create a service.}
\usage{
projects.locations.services.create(Service, parent)
}
\arguments{
\item{Service}{The \link{Service} object to pass to this method}

\item{parent}{The namespace in which the service should be created}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}

Other Service functions: \code{\link{Service}},
  \code{\link{namespaces.services.create}},
  \code{\link{namespaces.services.replaceService}},
  \code{\link{projects.locations.services.replaceService}}
}
\concept{Service functions}
