% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{projects.locations.services.getIamPolicy}
\alias{projects.locations.services.getIamPolicy}
\title{Get the IAM Access Control policy currently in effect for the givenCloud Run service. This result does not include any inherited policies.}
\usage{
projects.locations.services.getIamPolicy(resource,
  options.requestedPolicyVersion = NULL)
}
\arguments{
\item{resource}{REQUIRED: The resource for which the policy is being requested}

\item{options.requestedPolicyVersion}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}
}
