% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{projects.locations.services.setIamPolicy}
\alias{projects.locations.services.setIamPolicy}
\title{Sets the IAM Access control policy for the specified Service. Overwritesany existing policy.}
\usage{
projects.locations.services.setIamPolicy(SetIamPolicyRequest, resource)
}
\arguments{
\item{SetIamPolicyRequest}{The \link{SetIamPolicyRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy is being specified}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}
}
\concept{SetIamPolicyRequest functions}
