% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_run.R
\name{projects.locations.services.testIamPermissions}
\alias{projects.locations.services.testIamPermissions}
\title{Returns permissions that a caller has on the specified Project.There are no permissions required for making this API call.}
\usage{
projects.locations.services.testIamPermissions(TestIamPermissionsRequest,
  resource)
}
\arguments{
\item{TestIamPermissionsRequest}{The \link{TestIamPermissionsRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy detail is being requested}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/run/}{Google Documentation}

Other TestIamPermissionsRequest functions: \code{\link{TestIamPermissionsRequest}}
}
\concept{TestIamPermissionsRequest functions}
