% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_build.R
\name{projects.triggers.create}
\alias{projects.triggers.create}
\title{Creates a new `BuildTrigger`.This API is experimental.}
\usage{
projects.triggers.create(BuildTrigger, projectId)
}
\arguments{
\item{BuildTrigger}{The \link{BuildTrigger} object to pass to this method}

\item{projectId}{ID of the project for which to configure automatic builds}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/cloud-build/docs/}{Google Documentation}

Other BuildTrigger functions: \code{\link{projects.triggers.patch}}
}
\concept{BuildTrigger functions}
