% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_build.R
\name{projects.triggers.list}
\alias{projects.triggers.list}
\title{Lists existing `BuildTrigger`s.This API is experimental.}
\usage{
projects.triggers.list(projectId, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{projectId}{ID of the project for which to list BuildTriggers}

\item{pageToken}{Token to provide to skip to a particular spot in the list}

\item{pageSize}{Number of results to return in the list}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/cloud-build/docs/}{Google Documentation}
}
