% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_build.R
\name{projects.triggers.run}
\alias{projects.triggers.run}
\title{Runs a `BuildTrigger` at a particular source revision.}
\usage{
projects.triggers.run(RepoSource, projectId, triggerId)
}
\arguments{
\item{RepoSource}{The \link{RepoSource} object to pass to this method}

\item{projectId}{ID of the project}

\item{triggerId}{ID of the trigger}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/cloud-build/docs/}{Google Documentation}

Other RepoSource functions: \code{\link{RepoSource}}
}
\concept{RepoSource functions}
