% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtrigger_objects.R
\name{PubsubConfig}
\alias{PubsubConfig}
\title{Pubsub Config (Build Trigger)}
\usage{
PubsubConfig(
  subscription = NULL,
  topic = NULL,
  serviceAccountEmail = NULL,
  state = NULL
)
}
\arguments{
\item{subscription}{Output only. Name of the subscription.}

\item{topic}{The name of the topic from which this subscription is receiving messages.}

\item{serviceAccountEmail}{Service account that will make the push request.}

\item{state}{Potential issues with the underlying Pub/Sub subscription configuration. Only populated on get requests.}
}
\value{
A PubsubConfig object
}
\description{
PubsubConfig describes the configuration of a trigger that creates a build whenever a Pub/Sub message is published.
}
\seealso{
`https://cloud.google.com/build/docs/api/reference/rest/v1/projects.locations.triggers#BuildTrigger.PubsubConfig`
}
