% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{PubsubMessage.attributes}
\alias{PubsubMessage.attributes}
\title{PubsubMessage.attributes Object}
\usage{
PubsubMessage.attributes()
}
\value{
PubsubMessage.attributes object
}
\description{
PubsubMessage.attributes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional attributes for this message.
}
\seealso{
Other PubsubMessage functions: \code{\link{PubsubMessage}}
}
\concept{PubsubMessage functions}
