% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{PubsubMessage}
\alias{PubsubMessage}
\title{PubsubMessage Object}
\usage{
PubsubMessage(PubsubMessage.attributes = NULL, attributes = NULL,
  messageId = NULL, publishTime = NULL, data = NULL)
}
\arguments{
\item{PubsubMessage.attributes}{The \link{PubsubMessage.attributes} object or list of objects}

\item{attributes}{Optional attributes for this message}

\item{messageId}{ID of this message, assigned by the server when the message is published}

\item{publishTime}{The time at which the message was published, populated by the server when}

\item{data}{The message data field}
}
\value{
PubsubMessage object
}
\description{
PubsubMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message that is published by publishers and consumed by subscribers. Themessage must contain either a non-empty data field or at least one attribute.Note that client libraries represent this object differentlydepending on the language. See the corresponding<a href='https://cloud.google.com/pubsub/docs/reference/libraries'>clientlibrary documentation</a> for more information. See<a href='https://cloud.google.com/pubsub/quotas'>Quotas and limits</a>for more information about message limits.
}
\seealso{
Other PubsubMessage functions: \code{\link{PubsubMessage.attributes}}
}
\concept{PubsubMessage functions}
