% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{PubsubTarget.attributes}
\alias{PubsubTarget.attributes}
\title{PubsubTarget.attributes Object}
\usage{
PubsubTarget.attributes()
}
\value{
PubsubTarget.attributes object
}
\description{
PubsubTarget.attributes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Attributes for PubsubMessage.Pubsub message must contain either non-empty data, or at least oneattribute.
}
\seealso{
Other PubsubTarget functions: \code{\link{PubsubTarget}}
}
\concept{PubsubTarget functions}
