% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{PubsubTarget}
\alias{PubsubTarget}
\title{PubsubTarget Object}
\usage{
PubsubTarget(PubsubTarget.attributes = NULL, data = NULL,
  attributes = NULL, topicName = NULL)
}
\arguments{
\item{PubsubTarget.attributes}{The \link{PubsubTarget.attributes} object or list of objects}

\item{data}{The message payload for PubsubMessage}

\item{attributes}{Attributes for PubsubMessage}

\item{topicName}{Required}
}
\value{
PubsubTarget object
}
\description{
PubsubTarget Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Pub/Sub target. The job will be delivered by publishing a message tothe given Pub/Sub topic.
}
\seealso{
Other PubsubTarget functions: \code{\link{PubsubTarget.attributes}}
}
\concept{PubsubTarget functions}
