% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{PullRequestFilter}
\alias{PullRequestFilter}
\title{Cloud Build API Objects
Creates and manages builds on Google Cloud Platform.}
\usage{
PullRequestFilter(commentControl = NULL, branch = NULL)
}
\arguments{
\item{commentControl}{Whether to block builds on a '/gcbrun' comment from a repository owner or}

\item{branch}{Regex of branches to match}
}
\value{
PullRequestFilter object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2019-11-03 21:20:58
filename: R/objects_build.R
api_json: build
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
PullRequestFilter Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
PullRequestFilter contains filter properties for matching GitHub PullRequests.
}
\concept{PullRequestFilter functions}
