% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{PushFilter}
\alias{PushFilter}
\title{PushFilter Object}
\usage{
PushFilter(branch = NULL, tag = NULL)
}
\arguments{
\item{branch}{Regexes matching branches to build}

\item{tag}{Regexes matching tags to build}
}
\value{
PushFilter object
}
\description{
PushFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Push contains filter properties for matching GitHub git pushes.
}
\concept{PushFilter functions}
