% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild.R
\name{RepoSource}
\alias{RepoSource}
\title{RepoSource Object}
\usage{
RepoSource(repoName = NULL, tagName = NULL, projectId = NULL,
  commitSha = NULL, branchName = NULL, dir = NULL)
}
\arguments{
\item{repoName}{Name of the Cloud Source Repository}

\item{tagName}{Regex matching tags to build}

\item{projectId}{ID of the project that owns the Cloud Source Repository}

\item{commitSha}{Explicit commit SHA to build}

\item{branchName}{Regex matching branches to build e.g. ".*"}

\item{dir}{Directory, relative to the source root, in which to run the build}
}
\value{
RepoSource object
}
\description{
RepoSource Object
}
\details{
Location of the source in a Google Cloud Source Repository.

Only one of commitSha, branchName or tagName are allowed.
}
\examples{

\dontrun{

my_repo_source <- Source(repoSource=RepoSource("https://my-repo.com", branchName="master"))

build2 <- cr_build("cloudbuild.yaml", source = my_repo_source)

}
}
\seealso{
Other RepoSource functions: \code{\link{projects.triggers.run}}
}
\concept{RepoSource functions}
