% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_targets.R
\name{resolve_bucket_folder}
\alias{resolve_bucket_folder}
\title{Splits a target workflow across Cloud Build jobs}
\usage{
resolve_bucket_folder(target_folder, bucket)
}
\description{
Splits a target workflow across Cloud Build jobs
}
