% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ResourceRecord}
\alias{ResourceRecord}
\title{ResourceRecord Object}
\usage{
ResourceRecord(name = NULL, type = NULL, rrdata = NULL)
}
\arguments{
\item{name}{Relative name of the object affected by this record}

\item{type}{Resource record type}

\item{rrdata}{Data for this record}
}
\value{
ResourceRecord object
}
\description{
ResourceRecord Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A DNS resource record.
}
\concept{ResourceRecord functions}
