% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ResourceRequirements.limits}
\alias{ResourceRequirements.limits}
\title{ResourceRequirements.limits Object}
\usage{
ResourceRequirements.limits()
}
\value{
ResourceRequirements.limits object
}
\description{
ResourceRequirements.limits Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
(Optional)Cloud Run fully managed: Only memory and CPU are supported. Note: The onlysupported value for CPU is '1'.Cloud Run for Anthos: supportedLimits describes the maximum amount of compute resources allowed.The values of the map is string form of the 'quantity' k8s type:https://github.com/kubernetes/kubernetes/blob/master/staging/src/k8s.io/apimachinery/pkg/api/resource/quantity.go
}
\seealso{
Other ResourceRequirements functions: \code{\link{ResourceRequirements.requests}},
  \code{\link{ResourceRequirements}}
}
\concept{ResourceRequirements functions}
