% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ResourceRequirements}
\alias{ResourceRequirements}
\title{ResourceRequirements Object}
\usage{
ResourceRequirements(ResourceRequirements.limits = NULL,
  ResourceRequirements.requests = NULL, limits = NULL,
  requests = NULL)
}
\arguments{
\item{ResourceRequirements.limits}{The \link{ResourceRequirements.limits} object or list of objects}

\item{ResourceRequirements.requests}{The \link{ResourceRequirements.requests} object or list of objects}

\item{limits}{(Optional)}

\item{requests}{(Optional)}
}
\value{
ResourceRequirements object
}
\description{
ResourceRequirements Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ResourceRequirements describes the compute resource requirements.
}
\seealso{
Other ResourceRequirements functions: \code{\link{ResourceRequirements.limits}},
  \code{\link{ResourceRequirements.requests}}
}
\concept{ResourceRequirements functions}
