% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ResourceRequirements.requests}
\alias{ResourceRequirements.requests}
\title{ResourceRequirements.requests Object}
\usage{
ResourceRequirements.requests()
}
\value{
ResourceRequirements.requests object
}
\description{
ResourceRequirements.requests Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
(Optional)Cloud Run fully managed: Only memory and CPU are supported. Note: The onlysupported value for CPU is '1'.Cloud Run for Anthos: supportedRequests describes the minimum amount of compute resources required.If Requests is omitted for a container, it defaults to Limits if that isexplicitly specified, otherwise to an implementation-defined value.The values of the map is string form of the 'quantity' k8s type:https://github.com/kubernetes/kubernetes/blob/master/staging/src/k8s.io/apimachinery/pkg/api/resource/quantity.go
}
\seealso{
Other ResourceRequirements functions: \code{\link{ResourceRequirements.limits}},
  \code{\link{ResourceRequirements}}
}
\concept{ResourceRequirements functions}
