% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{Results}
\alias{Results}
\title{Results Object}
\usage{
Results(artifactManifest = NULL, buildStepImages = NULL,
  artifactTiming = NULL, buildStepOutputs = NULL, images = NULL,
  numArtifacts = NULL)
}
\arguments{
\item{artifactManifest}{Path to the artifact manifest}

\item{buildStepImages}{List of build step digests, in the order corresponding to build step}

\item{artifactTiming}{Time to push all non-container artifacts}

\item{buildStepOutputs}{List of build step outputs, produced by builder images, in the order}

\item{images}{Container images that were built as a part of the build}

\item{numArtifacts}{Number of artifacts uploaded}
}
\value{
Results object
}
\description{
Results Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Artifacts created by the build pipeline.
}
\concept{Results functions}
