% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{ResumeJobRequest}
\alias{ResumeJobRequest}
\title{ResumeJobRequest Object}
\usage{
ResumeJobRequest()
}
\value{
ResumeJobRequest object
}
\description{
ResumeJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for ResumeJob.
}
\seealso{
Other ResumeJobRequest functions: \code{\link{projects.locations.jobs.resume}}
}
\concept{ResumeJobRequest functions}
