% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_schedule.R
\name{RetryConfig}
\alias{RetryConfig}
\title{RetryConfig Object}
\usage{
RetryConfig(maxRetryDuration = NULL, retryCount = NULL,
  maxDoublings = NULL, minBackoffDuration = NULL,
  maxBackoffDuration = NULL)
}
\arguments{
\item{maxRetryDuration}{The time limit for retrying a failed job, measured from time when an}

\item{retryCount}{The number of attempts that the system will make to run a job using the}

\item{maxDoublings}{The time between retries will double `max_doublings` times}

\item{minBackoffDuration}{The minimum amount of time to wait before retrying a job after}

\item{maxBackoffDuration}{The maximum amount of time to wait before retrying a job after}
}
\value{
RetryConfig object
}
\description{
RetryConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings that determine the retry behavior.By default, if a job does not complete successfully (meaning thatan acknowledgement is not received from the handler, then it will be retriedwith exponential backoff according to the settings in RetryConfig.
}
\concept{RetryConfig functions}
