% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{Revision}
\alias{Revision}
\title{Revision Object}
\usage{
Revision(status = NULL, apiVersion = NULL, spec = NULL,
  metadata = NULL)
}
\arguments{
\item{status}{Status communicates the observed state of the Revision (from the}

\item{apiVersion}{The API version for this call such as 'serving}

\item{spec}{Spec holds the desired state of the Revision (from the client)}

\item{metadata}{Metadata associated with this Revision, including name, namespace, labels,}
}
\value{
Revision object
}
\description{
Revision Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Revision is an immutable snapshot of code and configuration.  A revisionreferences a container image. Revisions are created by updates to aConfiguration.See also:https://github.com/knative/serving/blob/master/docs/spec/overview.md#revision
}
\concept{Revision functions}
