% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudrun.R
\name{RevisionSpec}
\alias{RevisionSpec}
\title{RevisionSpec Object}
\usage{
RevisionSpec(timeoutSeconds = NULL, serviceAccountName = NULL,
  containers = NULL, volumes = NULL, containerConcurrency = NULL)
}
\arguments{
\item{timeoutSeconds}{TimeoutSeconds holds the max duration the instance is allowed for}

\item{serviceAccountName}{Email address of the IAM service account associated with the revision}

\item{containers}{Containers holds the single container that defines the unit of execution}

\item{volumes}{No description}

\item{containerConcurrency}{(Optional)}
}
\value{
RevisionSpec object
}
\description{
RevisionSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
RevisionSpec holds the desired state of the Revision (from the client).
}
\concept{RevisionSpec functions}
