% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RevisionStatus}
\alias{RevisionStatus}
\title{RevisionStatus Object}
\usage{
RevisionStatus(conditions = NULL, observedGeneration = NULL,
  imageDigest = NULL, logUrl = NULL, serviceName = NULL)
}
\arguments{
\item{conditions}{Conditions communicates information about ongoing/complete}

\item{observedGeneration}{ObservedGeneration is the 'Generation' of the Revision that}

\item{imageDigest}{ImageDigest holds the resolved digest for the image specified}

\item{logUrl}{Specifies the generated logging url for this particular revision}

\item{serviceName}{Not currently used by Cloud Run}
}
\value{
RevisionStatus object
}
\description{
RevisionStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
RevisionStatus communicates the observed state of the Revision (from thecontroller).
}
\concept{RevisionStatus functions}
