% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{RevisionTemplate}
\alias{RevisionTemplate}
\title{RevisionTemplate Object}
\usage{
RevisionTemplate(metadata = NULL, spec = NULL)
}
\arguments{
\item{metadata}{Optional metadata for this Revision, including labels and annotations}

\item{spec}{RevisionSpec holds the desired state of the Revision (from the client)}
}
\value{
RevisionTemplate object
}
\description{
RevisionTemplate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
RevisionTemplateSpec describes the data a revision should have when createdfrom a template. Based on:https://github.com/kubernetes/api/blob/e771f807/core/v1/types.go#L3179-L3190
}
\concept{RevisionTemplate functions}
