% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_build.R, R/funcs_run.R,
%   R/funcs_schedule.R
\name{rmNullObs}
\alias{rmNullObs}
\title{Recursively step down into list, removing all such objects}
\usage{
rmNullObs(x)

rmNullObs(x)

rmNullObs(x)
}
\description{
Recursively step down into list, removing all such objects

Recursively step down into list, removing all such objects

Recursively step down into list, removing all such objects
}
\keyword{internal}
