% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Route}
\alias{Route}
\title{Route Object}
\usage{
Route(metadata = NULL, status = NULL, apiVersion = NULL,
  spec = NULL)
}
\arguments{
\item{metadata}{Metadata associated with this Route, including name, namespace, labels,}

\item{status}{Status communicates the observed state of the Route (from the controller)}

\item{apiVersion}{The API version for this call such as 'serving}

\item{spec}{Spec holds the desired state of the Route (from the client)}
}
\value{
Route object
}
\description{
Route Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Route is responsible for configuring ingress over a collection of Revisions.Some of the Revisions a Route distributes traffic over may be specified byreferencing the Configuration responsible for creating them; in these casesthe Route is additionally responsible for monitoring the Configuration for'latest ready' revision changes, and smoothly rolling out latest revisions.See also:https://github.com/knative/serving/blob/master/docs/spec/overview.md#routeCloud Run currently supports referencing a single Configuration toautomatically deploy the 'latest ready' Revision from that Configuration.
}
\seealso{
Other Route functions: \code{\link{namespaces.routes.create}},
  \code{\link{namespaces.routes.replaceRoute}},
  \code{\link{projects.locations.routes.create}},
  \code{\link{projects.locations.routes.replaceRoute}}
}
\concept{Route functions}
