% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RouteSpec}
\alias{RouteSpec}
\title{RouteSpec Object}
\usage{
RouteSpec(traffic = NULL)
}
\arguments{
\item{traffic}{Traffic specifies how to distribute traffic over a collection of Knative}
}
\value{
RouteSpec object
}
\description{
RouteSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
RouteSpec holds the desired state of the Route (from the client).
}
\concept{RouteSpec functions}
