% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{RouteStatus}
\alias{RouteStatus}
\title{RouteStatus Object}
\usage{
RouteStatus(url = NULL, traffic = NULL, conditions = NULL,
  observedGeneration = NULL, address = NULL)
}
\arguments{
\item{url}{URL holds the url that will distribute traffic over the provided traffic}

\item{traffic}{Traffic holds the configured traffic distribution}

\item{conditions}{Conditions communicates information about ongoing/complete}

\item{observedGeneration}{ObservedGeneration is the 'Generation' of the Route that}

\item{address}{Similar to url, information on where the service is available on HTTP}
}
\value{
RouteStatus object
}
\description{
RouteStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
RouteStatus communicates the observed state of the Route (from thecontroller).
}
\concept{RouteStatus functions}
