% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\docType{package}
\name{run_googleAuthR}
\alias{run_googleAuthR}
\alias{run_googleAuthR-package}
\title{Cloud Run API
Deploy and manage user provided container images that scale automatically based on HTTP traffic.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2019-11-03 11:19:16
filename: R/run.R
api_json: cr
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
