% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{Secret}
\alias{Secret}
\title{Secret Object}
\usage{
Secret(Secret.secretEnv = NULL, kmsKeyName = NULL, secretEnv = NULL)
}
\arguments{
\item{Secret.secretEnv}{The \link{Secret.secretEnv} object or list of objects}

\item{kmsKeyName}{Cloud KMS key name to use to decrypt these envs}

\item{secretEnv}{Map of environment variable name to its encrypted value}
}
\value{
Secret object
}
\description{
Secret Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Pairs a set of secret environment variables containing encryptedvalues with the Cloud KMS key to use to decrypt the value.
}
\seealso{
Other Secret functions: \code{\link{Secret.secretEnv}}
}
\concept{Secret functions}
