% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{Secret.secretEnv}
\alias{Secret.secretEnv}
\title{Secret.secretEnv Object}
\usage{
Secret.secretEnv()
}
\value{
Secret.secretEnv object
}
\description{
Secret.secretEnv Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Map of environment variable name to its encrypted value.Secret environment variables must be unique across all of a build'ssecrets, and must be used by at least one build step. Values can be at most64 KB in size. There can be at most 100 secret values across all of abuild's secrets.
}
\seealso{
Other Secret functions: \code{\link{Secret}}
}
\concept{Secret functions}
