% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{SecretEnvSource}
\alias{SecretEnvSource}
\title{SecretEnvSource Object}
\usage{
SecretEnvSource(name = NULL, localObjectReference = NULL,
  optional = NULL)
}
\arguments{
\item{name}{Cloud Run fully managed: not supported}

\item{localObjectReference}{This field should not be used directly as it is meant to be inlined}

\item{optional}{(Optional)}
}
\value{
SecretEnvSource object
}
\description{
SecretEnvSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedSecretEnvSource selects a Secret to populate the environmentvariables with.The contents of the target Secret's Data field will represent thekey-value pairs as environment variables.
}
\concept{SecretEnvSource functions}
