% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{SecretKeySelector}
\alias{SecretKeySelector}
\title{SecretKeySelector Object}
\usage{
SecretKeySelector(key = NULL, name = NULL,
  localObjectReference = NULL, optional = NULL)
}
\arguments{
\item{key}{Cloud Run fully managed: not supported}

\item{name}{Cloud Run fully managed: not supported}

\item{localObjectReference}{This field should not be used directly as it is meant to be inlined}

\item{optional}{(Optional)}
}
\value{
SecretKeySelector object
}
\description{
SecretKeySelector Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedSecretKeySelector selects a key of a Secret.
}
\concept{SecretKeySelector functions}
