% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SecretVolumeSource}
\alias{SecretVolumeSource}
\title{SecretVolumeSource Object}
\usage{
SecretVolumeSource(defaultMode = NULL, secretName = NULL,
  items = NULL, optional = NULL)
}
\arguments{
\item{defaultMode}{(Optional)}

\item{secretName}{Cloud Run fully managed: not supported}

\item{items}{(Optional)}

\item{optional}{(Optional)}
}
\value{
SecretVolumeSource object
}
\description{
SecretVolumeSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedThe contents of the target Secret's Data field will be presented in a volumeas files using the keys in the Data field as the file names.
}
\concept{SecretVolumeSource functions}
