% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SecurityContext}
\alias{SecurityContext}
\title{SecurityContext Object}
\usage{
SecurityContext(runAsUser = NULL)
}
\arguments{
\item{runAsUser}{(Optional)}
}
\value{
SecurityContext object
}
\description{
SecurityContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedSecurityContext holds security configuration that will be applied to acontainer. Some fields are present in both SecurityContext andPodSecurityContext.  When both are set, the values in SecurityContext takeprecedence.
}
\concept{SecurityContext functions}
