% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ServiceSpec}
\alias{ServiceSpec}
\title{ServiceSpec Object}
\usage{
ServiceSpec(traffic = NULL, template = NULL)
}
\arguments{
\item{traffic}{Traffic specifies how to distribute traffic over a collection of Knative}

\item{template}{Template holds the latest specification for the Revision to}
}
\value{
ServiceSpec object
}
\description{
ServiceSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ServiceSpec holds the desired state of the Route (from the client), whichis used to manipulate the underlying Route and Configuration(s).
}
\concept{ServiceSpec functions}
