% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{ServiceStatus}
\alias{ServiceStatus}
\title{ServiceStatus Object}
\usage{
ServiceStatus(conditions = NULL, address = NULL,
  observedGeneration = NULL, url = NULL, traffic = NULL,
  latestReadyRevisionName = NULL, latestCreatedRevisionName = NULL)
}
\arguments{
\item{conditions}{Conditions communicates information about ongoing/complete}

\item{address}{From RouteStatus}

\item{observedGeneration}{ObservedGeneration is the 'Generation' of the Route that}

\item{url}{From RouteStatus}

\item{traffic}{From RouteStatus}

\item{latestReadyRevisionName}{From ConfigurationStatus}

\item{latestCreatedRevisionName}{From ConfigurationStatus}
}
\value{
ServiceStatus object
}
\description{
ServiceStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The current state of the Service. Output only.
}
\concept{ServiceStatus functions}
