% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{SetIamPolicyRequest}
\alias{SetIamPolicyRequest}
\title{SetIamPolicyRequest Object}
\usage{
SetIamPolicyRequest(policy = NULL, updateMask = NULL)
}
\arguments{
\item{policy}{REQUIRED: The complete policy to be applied to the `resource`}

\item{updateMask}{OPTIONAL: A FieldMask specifying which fields of the policy to modify}
}
\value{
SetIamPolicyRequest object
}
\description{
SetIamPolicyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for `SetIamPolicy` method.
}
\seealso{
Other SetIamPolicyRequest functions: \code{\link{projects.locations.services.setIamPolicy}}
}
\concept{SetIamPolicyRequest functions}
