% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{SourceProvenance.fileHashes}
\alias{SourceProvenance.fileHashes}
\title{SourceProvenance.fileHashes Object}
\usage{
SourceProvenance.fileHashes()
}
\value{
SourceProvenance.fileHashes object
}
\description{
SourceProvenance.fileHashes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output only. Hash(es) of the build source, which can be used to verify thatthe original source integrity was maintained in the build. Note that`FileHashes` will only be populated if `BuildOptions` has requested a`SourceProvenanceHash`.The keys to this map are file paths used as build source and the valuescontain the hash values for those files.If the build source came in a single package such as a gzipped tarfile(`.tar.gz`), the `FileHash` will be for the single path to that file.
}
\seealso{
Other SourceProvenance functions: \code{\link{SourceProvenance}}
}
\concept{SourceProvenance functions}
