% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{StatusCause}
\alias{StatusCause}
\title{StatusCause Object}
\usage{
StatusCause(message = NULL, reason = NULL, field = NULL)
}
\arguments{
\item{message}{A human-readable description of the cause of the error}

\item{reason}{A machine-readable description of the cause of the error}

\item{field}{The field of the resource that has caused this error, as named by its JSON}
}
\value{
StatusCause object
}
\description{
StatusCause Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
StatusCause provides more information about an api.Status failure, includingcases when multiple errors are encountered.
}
\concept{StatusCause functions}
