% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{StatusDetails}
\alias{StatusDetails}
\title{StatusDetails Object}
\usage{
StatusDetails(name = NULL, uid = NULL, group = NULL, causes = NULL,
  retryAfterSeconds = NULL)
}
\arguments{
\item{name}{The name attribute of the resource associated with the status StatusReason}

\item{uid}{UID of the resource}

\item{group}{The group attribute of the resource associated with the status}

\item{causes}{The Causes array includes more details associated with the StatusReason}

\item{retryAfterSeconds}{If specified, the time in seconds before the operation should be retried}
}
\value{
StatusDetails object
}
\description{
StatusDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
StatusDetails is a set of additional properties that MAY be set by theserver to provide additional information about a response. The Reasonfield of a Status object defines what attributes will be set. Clientsmust ignore fields that do not match the defined type of each attribute,and should assume that any attribute may be empty, invalid, or underdefined.
}
\concept{StatusDetails functions}
