% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{StorageSource}
\alias{StorageSource}
\title{StorageSource Object}
\usage{
StorageSource(bucket = NULL, object = NULL, generation = NULL)
}
\arguments{
\item{bucket}{Google Cloud Storage bucket containing the source (see}

\item{object}{Google Cloud Storage object containing the source}

\item{generation}{Google Cloud Storage generation for the object}
}
\value{
StorageSource object
}
\description{
StorageSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Location of the source in an archive file in Google Cloud Storage.
}
\concept{StorageSource functions}
