% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{TCPSocketAction}
\alias{TCPSocketAction}
\title{TCPSocketAction Object}
\usage{
TCPSocketAction(port = NULL, host = NULL)
}
\arguments{
\item{port}{Cloud Run fully managed: not supported}

\item{host}{(Optional)}
}
\value{
TCPSocketAction object
}
\description{
TCPSocketAction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedTCPSocketAction describes an action based on opening a socket
}
\concept{TCPSocketAction functions}
