% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R
\name{TestIamPermissionsRequest}
\alias{TestIamPermissionsRequest}
\title{TestIamPermissionsRequest Object}
\usage{
TestIamPermissionsRequest(permissions = NULL)
}
\arguments{
\item{permissions}{The set of permissions to check for the `resource`}
}
\value{
TestIamPermissionsRequest object
}
\description{
TestIamPermissionsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for `TestIamPermissions` method.
}
\seealso{
Other TestIamPermissionsRequest functions: \code{\link{projects.locations.services.testIamPermissions}}
}
\concept{TestIamPermissionsRequest functions}
