% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{TimeSpan}
\alias{TimeSpan}
\title{TimeSpan Object}
\usage{
TimeSpan(endTime = NULL, startTime = NULL)
}
\arguments{
\item{endTime}{End of time span}

\item{startTime}{Start of time span}
}
\value{
TimeSpan object
}
\description{
TimeSpan Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Start and end times for a build execution phase.
}
\concept{TimeSpan functions}
