% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{TrafficTarget}
\alias{TrafficTarget}
\title{TrafficTarget Object}
\usage{
TrafficTarget(tag = NULL, url = NULL, configurationName = NULL,
  percent = NULL, revisionName = NULL, latestRevision = NULL)
}
\arguments{
\item{tag}{Tag is optionally used to expose a dedicated url for referencing}

\item{url}{Output only}

\item{configurationName}{ConfigurationName of a configuration to whose latest revision we will}

\item{percent}{Percent specifies percent of the traffic to this Revision or Configuration}

\item{revisionName}{RevisionName of a specific revision to which to send this portion of}

\item{latestRevision}{LatestRevision may be optionally provided to indicate that the latest}
}
\value{
TrafficTarget object
}
\description{
TrafficTarget Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
TrafficTarget holds a single entry of the routing table for a Route.
}
\concept{TrafficTarget functions}
