% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R, R/objects_run.R
\name{Volume}
\alias{Volume}
\title{Volume Object}
\usage{
Volume(name = NULL, secret = NULL, configMap = NULL)

Volume(name = NULL, secret = NULL, configMap = NULL)
}
\arguments{
\item{name}{Name of the volume to mount}

\item{secret}{Cloud Run fully managed: not supported}

\item{configMap}{Cloud Run fully managed: not supported}

\item{path}{Path at which to mount the volume}

\item{name}{Cloud Run fully managed: not supported}
}
\value{
Volume object

Volume object
}
\description{
Volume Object

Volume Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Volume describes a Docker container volume which is mounted into build stepsin order to persist files across build step execution.

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedVolume represents a named volume in a container.
}
\concept{Volume functions}
