% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtrigger_objects.R
\name{WebhookConfig}
\alias{WebhookConfig}
\title{WebhookConfig (Build Triggers)}
\usage{
WebhookConfig(secret, state = NULL)
}
\arguments{
\item{secret}{Resource name for the secret required as a URL parameter.}

\item{state}{Potential issues with the underlying Pub/Sub subscription configuration. Only populated on get requests.}
}
\value{
A WebhookConfig object
}
\description{
WebhookConfig describes the configuration of a trigger that creates a build whenever a webhook is sent to a trigger's webhook URL.
}
