% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{cr_bucket_set}
\alias{cr_bucket_set}
\alias{cr_bucket_get}
\title{Get/Set the Cloud Storage bucket for your Cloud Build Service}
\usage{
cr_bucket_set(bucket)

cr_bucket_get()
}
\arguments{
\item{bucket}{The GCS bucket}
}
\description{
Can also use environment arg GCS_DEFAULT_BUCKET
}
\examples{

cr_bucket_get()
}
