% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_schedule.R
\name{cr_build_schedule_http}
\alias{cr_build_schedule_http}
\title{Create a Cloud Scheduler HTTP target from a Cloud Build object}
\usage{
cr_build_schedule_http(build, projectId = cr_project_get())
}
\arguments{
\item{build}{A \link{Build} object}

\item{projectId}{The projectId}
}
\value{
A \link{HttpTarget} object for use in \link{cr_schedule}
}
\description{
This enables Cloud Scheduler to trigger Cloud Builds
}
\seealso{
https://cloud.google.com/cloud-build/docs/api/reference/rest/v1/projects.builds/create
}
\concept{Cloud Scheduler functions, Cloud Build functions}
