% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_schedule.R
\name{cr_build_schedule_http}
\alias{cr_build_schedule_http}
\title{Create a Cloud Scheduler HTTP target from a Cloud Build object}
\usage{
cr_build_schedule_http(build, projectId = cr_project_get())
}
\arguments{
\item{build}{A \link{Build} object usually created with \link{cr_build_make}}

\item{projectId}{The projectId}
}
\value{
A \link{HttpTarget} object for use in \link{cr_schedule}
}
\description{
This enables Cloud Scheduler to trigger Cloud Builds
}
\details{
Ensure you have a service email with \link{cr_email_set} of format \code{service-{project-number}@gcp-sa-cloudscheduler.iam.gserviceaccount.com} with Cloud Scheduler Service Agent role as per https://cloud.google.com/scheduler/docs/http-target-auth#add
}
\examples{
cloudbuild <- system.file("cloudbuild/cloudbuild.yaml", package = "googleCloudRunner")
build1 <- cr_build_make(cloudbuild)
build1
\dontrun{
cr_schedule("15 5 * * *", name="cloud-build-test1",
            httpTarget = cr_build_schedule_http(build1))
}

}
\seealso{
https://cloud.google.com/cloud-build/docs/api/reference/rest/v1/projects.builds/create
}
\concept{Cloud Scheduler functions, Cloud Build functions}
