% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_schedule.R
\name{cr_build_schedule_http}
\alias{cr_build_schedule_http}
\alias{cr_schedule_build}
\title{Create a Cloud Scheduler HTTP target from a Cloud Build object}
\usage{
cr_build_schedule_http(
  build,
  email = cr_email_get(),
  projectId = cr_project_get()
)

cr_schedule_build(
  build,
  schedule,
  schedule_type = c("http", "pubsub"),
  schedule_pubsub = NULL,
  email = cr_email_get(),
  projectId = cr_project_get(),
  ...
)
}
\arguments{
\item{build}{A \link{Build} object created via \link{cr_build_make} or \link{cr_build}}

\item{email}{The email that will authenticate the job set via \link{cr_email_set}}

\item{projectId}{The projectId}

\item{schedule}{A cron schedule e.g. \code{"15 5 * * *"}}

\item{schedule_type}{Whether to use HTTP or PubSub styled schedules}

\item{schedule_pubsub}{If you have a custom pubsub message to send via an existing topic, use \link{cr_schedule_pubsub} to supply it here}

\item{...}{
  Arguments passed on to \code{\link[=cr_schedule]{cr_schedule}}
  \describe{
    \item{\code{region}}{The region usually set with \link{cr_region_set}}
    \item{\code{overwrite}}{If TRUE and an existing job with the same name exists, will overwrite it with the new parameters}
    \item{\code{name}}{Name to call your scheduled job}
    \item{\code{httpTarget}}{A HTTP target object \link{HttpTarget}}
    \item{\code{pubsubTarget}}{A Pub/Sub target object \link{PubsubTarget} such as created via \link{cr_schedule_pubsub}}
    \item{\code{description}}{Optionally caller-specified in CreateJob or}
    \item{\code{timeZone}}{Specifies the time zone to be used in interpreting schedule. If set to \code{NULL} will be "UTC". Note that some time zones include a provision for daylight savings time.}
  }}
}
\value{
A \link{HttpTarget} object for use in \link{cr_schedule}

a \link{HttpTarget} object for use in \link{cr_schedule}

A cloud scheduler \link{Job} object
}
\description{
This enables Cloud Scheduler to trigger Cloud Builds
}
\details{
Ensure you have a service email with \link{cr_email_set} of format
\code{service-{project-number}@gcp-sa-cloudscheduler.iam.gserviceaccount.com}
with Cloud Scheduler Service Agent role as per
https://cloud.google.com/scheduler/docs/http-target-auth#add

See also \link{cr_schedule_pubsub} which you can use by creating
a build trigger of your build via \link{cr_buildtrigger} that accepts
Pub/Sub messages.  This method is recommended as being easier to maintain
than using HTTP requests to the Cloud Build API that
\link{cr_build_schedule_http} produces.
}
\examples{
cloudbuild <- system.file("cloudbuild/cloudbuild.yaml", package = "googleCloudRunner")
build1 <- cr_build_make(cloudbuild)
build1
\dontrun{
cr_schedule("cloud-build-test1",
  schedule = "15 5 * * *",
  httpTarget = cr_build_schedule_http(build1)
)

# a cloud build you would like to schedule
itworks <- cr_build("cloudbuild.yaml", launch_browser = FALSE)

# once working, pass in the build to the scheduler
cr_schedule("itworks-schedule",
  schedule = "15 5 * * *",
  httpTarget = cr_build_schedule_http(itworks)
)
}

}
\seealso{
https://cloud.google.com/build/docs/api/reference/rest/v1/projects.builds/create

Other Cloud Scheduler functions: 
\code{\link{HttpTarget}()},
\code{\link{Job}()},
\code{\link{PubsubTarget}()},
\code{\link{cr_run_schedule_http}()},
\code{\link{cr_schedule_delete}()},
\code{\link{cr_schedule_get}()},
\code{\link{cr_schedule_list}()},
\code{\link{cr_schedule_pause}()},
\code{\link{cr_schedule_pubsub}()},
\code{\link{cr_schedule_run}()},
\code{\link{cr_schedule}()}
}
\concept{Cloud Scheduler functions}
