% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudrun.R
\name{cr_build_step}
\alias{cr_build_step}
\title{Create a yaml build step}
\usage{
cr_build_step(name, args, stem = "gcr.io/cloud-builders/")
}
\arguments{
\item{name}{name of SDK appended to stem}

\item{args}{character vector of arguments}

\item{stem}{prefixed to name}
}
\description{
Create a yaml build step
}
