% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildsteps.R
\name{cr_buildstep_docker}
\alias{cr_buildstep_docker}
\title{Create a build step to build and push a docker image}
\usage{
cr_buildstep_docker(image, location = ".", dir = "")
}
\arguments{
\item{image}{The image tag that will be pushed}

\item{location}{Which folder to build}
}
\description{
Create a build step to build and push a docker image
}
\examples{
cr_buildstep_docker("gcr.io/my-project/my-image")
}
