% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{cr_buildstep_docker_auth}
\alias{cr_buildstep_docker_auth}
\title{Authorize Docker using \code{gcloud auth configure-docker}}
\usage{
cr_buildstep_docker_auth(image, ...)
}
\arguments{
\item{image}{name of the Docker image to push or pull from that needs
authorization, or simply the registry.}

\item{...}{Other arguments passed to \link{cr_buildstep_gcloud}}
}
\value{
A buildstep
}
\description{
Authorize Docker using \code{gcloud auth configure-docker}
}
\examples{
cr_buildstep_docker_auth("us.gcr.io")
cr_buildstep_docker_auth(c("us.gcr.io", "asia.gcr.io"))
cr_buildstep_docker_auth("https://asia.gcr.io/myrepo/image")
}
