% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildsteps_templates.R
\name{cr_buildstep_run}
\alias{cr_buildstep_run}
\title{Create buildsteps to deploy to Cloud Run}
\usage{
cr_buildstep_run(
  name,
  image,
  allowUnauthenticated = TRUE,
  region = cr_region_get(),
  concurrency = 80,
  ...
)
}
\arguments{
\item{name}{Name for deployment on Cloud Run}

\item{image}{The name of the image to create or use in deployment - \code{gcr.io}}

\item{allowUnauthenticated}{TRUE if can be reached from public HTTP address.}

\item{region}{The endpoint region for deployment}

\item{concurrency}{How many connections each image can serve. Can be up to 80.}

\item{...}{passed on to \link{cr_buildstep}}
}
\description{
Create buildsteps to deploy to Cloud Run
}
\seealso{
Other Cloud Buildsteps: 
\code{\link{cr_buildstep_bash}()},
\code{\link{cr_buildstep_decrypt}()},
\code{\link{cr_buildstep_df}()},
\code{\link{cr_buildstep_docker}()},
\code{\link{cr_buildstep_edit}()},
\code{\link{cr_buildstep_extract}()},
\code{\link{cr_buildstep_gitsetup}()},
\code{\link{cr_buildstep_mailgun}()},
\code{\link{cr_buildstep_nginx_setup}()},
\code{\link{cr_buildstep_pkgdown}()},
\code{\link{cr_buildstep_r}()},
\code{\link{cr_buildstep_slack}()},
\code{\link{cr_buildstep}()}
}
\concept{Cloud Buildsteps}
