% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger}
\alias{cr_buildtrigger}
\title{Creates a new `BuildTrigger`.This API is experimental.}
\usage{
cr_buildtrigger(name, trigger, build, description = NULL, tags = NULL,
  disabled = FALSE, substitutions = NULL, ignoredFiles = NULL,
  includedFiles = NULL, projectId = cr_project_get())
}
\arguments{
\item{name}{User assigned name of the trigger}

\item{trigger}{The trigger which will be a \link{RepoSource} or a \link{GitHubEventsConfig}}

\item{build}{A file location within the trigger source to use for the build steps, or a \link{Build} object}

\item{description}{Human-readable description of this trigger}

\item{tags}{Tags for annotation of a `BuildTrigger`}

\item{disabled}{If true, the trigger will never result in a build}

\item{substitutions}{A named list of Build macro variables}

\item{ignoredFiles}{ignored_files and included_files are file glob matches extended with support for "**".}

\item{includedFiles}{If any of the files altered in the commit pass the ignored_files}

\item{projectId}{ID of the project for which to configure automatic builds}
}
\description{
Creates a new `BuildTrigger`.This API is experimental.
}
\seealso{
Other BuildTrigger functions: \code{\link{BuildTrigger}},
  \code{\link{GitHubEventsConfig}},
  \code{\link{cr_buildtrigger_delete}},
  \code{\link{cr_buildtrigger_edit}},
  \code{\link{cr_buildtrigger_get}},
  \code{\link{cr_buildtrigger_list}},
  \code{\link{cr_buildtrigger_run}}
}
\concept{BuildTrigger functions}
