% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger_delete}
\alias{cr_buildtrigger_delete}
\title{Deletes a `BuildTrigger` by its project ID and trigger ID.This API is experimental.}
\usage{
cr_buildtrigger_delete(triggerId, projectId = cr_project_get())
}
\arguments{
\item{triggerId}{ID of the `BuildTrigger` to get or a \code{BuildTriggerResponse} object}

\item{projectId}{ID of the project that owns the trigger}
}
\description{
Deletes a `BuildTrigger` by its project ID and trigger ID.This API is experimental.
}
\seealso{
Other BuildTrigger functions: 
\code{\link{BuildTrigger}()},
\code{\link{GitHubEventsConfig}()},
\code{\link{cr_buildtrigger_edit}()},
\code{\link{cr_buildtrigger_get}()},
\code{\link{cr_buildtrigger_list}()},
\code{\link{cr_buildtrigger_make}()},
\code{\link{cr_buildtrigger_run}()},
\code{\link{cr_buildtrigger}()}
}
\concept{BuildTrigger functions}
