% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger_edit}
\alias{cr_buildtrigger_edit}
\title{Updates a `BuildTrigger` by its project ID and trigger ID.This API is experimental.}
\usage{
cr_buildtrigger_edit(BuildTrigger, projectId, triggerId)
}
\arguments{
\item{BuildTrigger}{The \link{BuildTrigger} object to pass to this method}

\item{projectId}{ID of the project that owns the trigger}

\item{triggerId}{ID of the `BuildTrigger` to update}
}
\description{
Updates a `BuildTrigger` by its project ID and trigger ID.This API is experimental.
}
\seealso{
Other BuildTrigger functions: \code{\link{BuildTrigger}},
  \code{\link{cr_buildtrigger}}
}
\concept{BuildTrigger functions}
