% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger_get}
\alias{cr_buildtrigger_get}
\title{Returns information about a `BuildTrigger`.This API is experimental.}
\usage{
cr_buildtrigger_get(triggerId, projectId = cr_project_get())
}
\arguments{
\item{triggerId}{ID of the `BuildTrigger` to get or a \code{BuildTriggerResponse} object}

\item{projectId}{ID of the project that owns the trigger}
}
\description{
Returns information about a `BuildTrigger`.This API is experimental.
}
\seealso{
Other BuildTrigger functions: 
\code{\link{BuildTrigger}()},
\code{\link{GitHubEventsConfig}()},
\code{\link{cr_buildtrigger_copy}()},
\code{\link{cr_buildtrigger_delete}()},
\code{\link{cr_buildtrigger_edit}()},
\code{\link{cr_buildtrigger_list}()},
\code{\link{cr_buildtrigger_repo}()},
\code{\link{cr_buildtrigger_run}()},
\code{\link{cr_buildtrigger}()}
}
\concept{BuildTrigger functions}
