% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger_list}
\alias{cr_buildtrigger_list}
\title{Lists existing `BuildTrigger`s.This API is experimental.}
\usage{
cr_buildtrigger_list(projectId = cr_project_get())
}
\arguments{
\item{projectId}{ID of the project for which to list BuildTriggers}
}
\description{
Lists existing `BuildTrigger`s.This API is experimental.
}
\seealso{
Other BuildTrigger functions: 
\code{\link{BuildTrigger}()},
\code{\link{GitHubEventsConfig}()},
\code{\link{cr_buildtrigger_delete}()},
\code{\link{cr_buildtrigger_edit}()},
\code{\link{cr_buildtrigger_get}()},
\code{\link{cr_buildtrigger_make}()},
\code{\link{cr_buildtrigger_run}()},
\code{\link{cr_buildtrigger}()}
}
\concept{BuildTrigger functions}
