% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtriggers.R
\name{cr_buildtrigger_list}
\alias{cr_buildtrigger_list}
\title{Lists existing `BuildTrigger`s.This API is experimental.}
\usage{
cr_buildtrigger_list(projectId = cr_project_get(), full = FALSE)
}
\arguments{
\item{projectId}{ID of the project for which to list BuildTriggers}

\item{full}{Return the trigger list with the build information from
\code{\link{cr_buildtrigger_get}}}
}
\description{
Lists existing `BuildTrigger`s.This API is experimental.
}
\examples{
\dontrun{

cr_buildtrigger_list()
}
}
\seealso{
\link{cr_build_list} which merges with this list

Other BuildTrigger functions: 
\code{\link{BuildTrigger}()},
\code{\link{GitHubEventsConfig}()},
\code{\link{cr_buildtrigger_copy}()},
\code{\link{cr_buildtrigger_delete}()},
\code{\link{cr_buildtrigger_edit}()},
\code{\link{cr_buildtrigger_get}()},
\code{\link{cr_buildtrigger_pubsub}()},
\code{\link{cr_buildtrigger_repo}()},
\code{\link{cr_buildtrigger_run}()},
\code{\link{cr_buildtrigger_webhook}()},
\code{\link{cr_buildtrigger}()}
}
\concept{BuildTrigger functions}
