% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{cr_deploy}
\alias{cr_deploy}
\title{Deploy an R script to Cloud}
\usage{
cr_deploy(local, remote = basename(local), dockerfile = NULL,
  image_name = remote, region = cr_region_get(),
  bucket = Sys.getenv("GCS_DEFAULT_BUCKET"),
  projectId = Sys.getenv("GCE_DEFAULT_PROJECT_ID"))
}
\arguments{
\item{local}{A folder containing the R script and all its R dependencies}

\item{remote}{The folder on Google Cloud Storage}

\item{dockerfile}{An optional Dockerfile built to support the script}

\item{image_name}{The gcr.io image name that will be deployed and/or built}

\item{region}{The Cloud Run endpoint set by CR_REGION env arg}

\item{bucket}{The Cloud Storage bucket that will hold the code}

\item{projectId}{The projectId where it all gets deployed to}
}
\description{
Helper to create the Dockerfile
}
\examples{

\dontrun{

cr_deploy("inst/example/", remote = "cloudRunnerTest")

}
}
