% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{cr_deploy_docker}
\alias{cr_deploy_docker}
\title{Deploy a local Dockerfile to be built on ContainerRegistry}
\usage{
cr_deploy_docker(
  local,
  image_name = remote,
  dockerfile = NULL,
  remote = basename(local),
  tag = "$BUILD_ID",
  timeout = 600L,
  bucket = cr_bucket_get(),
  projectId = cr_project_get(),
  launch_browser = interactive(),
  ...
)
}
\arguments{
\item{local}{The folder containing the Dockerfile to build}

\item{image_name}{The name of the docker image to be built either full name starting with gcr.io or constructed from the image_name and projectId via \code{gcr.io/{projectId}/{image_name}}}

\item{dockerfile}{An optional Dockerfile built to support the script.  Not needed if 'Dockerfile' exists in folder.  If supplied will be copied into deployment folder and called "Dockerfile"}

\item{remote}{The folder on Google Cloud Storage}

\item{tag}{The tag to attached to the pushed image - can use \code{Build} macros}

\item{timeout}{Amount of time that this build should be allowed to run, to second}

\item{bucket}{The GCS bucket that will be used to deploy code source}

\item{projectId}{The projectId}

\item{launch_browser}{Whether to launch the logs URL in a browser once deployed}

\item{...}{Other arguments passed to \link{cr_buildstep_docker}}
}
\description{
Build a local Dockerfile in the cloud. See googleCloudRunner website for help how to generate Dockerfiles.
}
\details{
This lets you deploy local folders with Dockerfiles, automating saving the source on Google Cloud Storage.

To deploy builds on git triggers and sources such as GitHub, see the examples of \link{cr_buildstep_docker} or the use cases on the website
}
\examples{

\dontrun{
cr_project_set("my-project")
cr_region_set("europe-west1")
cr_email_set("123456@projectid.iam.gserviceaccount.com")
cr_bucket_set("my-bucket")

b <- cr_deploy_docker(system.file("example/", package="googleCloudRunner"))

}
}
\seealso{
Other Deployment functions: 
\code{\link{cr_deploy_git_html}()},
\code{\link{cr_deploy_github_docker}()},
\code{\link{cr_deploy_packagetests}()},
\code{\link{cr_deploy_pkgdown}()},
\code{\link{cr_deploy_run}()},
\code{\link{cr_deploy_r}()}
}
\concept{Deployment functions}
