% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{cr_email_get}
\alias{cr_email_get}
\alias{cr_email_set}
\title{Get/Set cloud build email}
\usage{
cr_email_get()

cr_email_set(cloudbuildEmail)
}
\arguments{
\item{cloudbuildEmail}{The Cloud Build service email}
}
\description{
Needed so Cloud Scheduler can run Cloud Build jobs - can also set via environment argument CR_BUILD_EMAIL
}
\seealso{
https://console.cloud.google.com/cloud-build/settings
}
