% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{cr_region_set}
\alias{cr_region_set}
\alias{cr_region_get}
\title{Get/Set the endpoint for your CloudRun services}
\usage{
cr_region_set(region = googleCloudRunner::cr_regions)

cr_region_get()
}
\arguments{
\item{region}{Region for the endpoint}
}
\description{
Can also use environment argument CR_REGION
}
\examples{

cr_region_get()
}
