% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudrun.R
\name{cr_run_get}
\alias{cr_run_get}
\title{Get information about a Cloud Run service.}
\usage{
cr_run_get(name, projectId = cr_project_get())
}
\arguments{
\item{name}{The name of the service to retrieve}

\item{projectId}{The projectId to get from}
}
\description{
Get information about a Cloud Run service.
}
\details{
This returns details on a particular deployed Cloud Run service.
}
\seealso{
\href{https://cloud.google.com/run/docs/reference/rest/v1/namespaces.services/get}{Google Documentation on namespaces.services.get}

Other Cloud Run functions: 
\code{\link{cr_jwt_create}()},
\code{\link{cr_plumber_pubsub}()},
\code{\link{cr_run_list}()},
\code{\link{cr_run}()}
}
\concept{Cloud Run functions}
