% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudrun.R
\name{cr_run_list}
\alias{cr_run_list}
\title{List CloudRun services.}
\usage{
cr_run_list(
  projectId = cr_project_get(),
  labelSelector = NULL,
  limit = NULL,
  summary = TRUE
)
}
\arguments{
\item{projectId}{The GCP project from which the services should be listed}

\item{labelSelector}{Allows to filter resources based on a label}

\item{limit}{The maximum number of records that should be returned}

\item{summary}{If TRUE will return only a subset of info available, set to FALSE for all metadata}
}
\description{
List the Cloud Run services you have access to
}
\seealso{
\href{https://cloud.run}{Google Documentation for Cloud Run}

Other Cloud Run functions: 
\code{\link{cr_plumber_pubsub}()},
\code{\link{cr_run_get}()},
\code{\link{cr_run}()}
}
\concept{Cloud Run functions}
