% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudscheduler.R
\name{cr_schedule}
\alias{cr_schedule}
\title{Creates a Cloud Scheduler job.}
\usage{
cr_schedule(schedule, name = NULL, httpTarget = NULL,
  description = NULL, projectId = Sys.getenv("GCE_DEFAULT_PROJECT_ID"))
}
\arguments{
\item{schedule}{Required, except when used with UpdateJob}

\item{name}{Optionally caller-specified in CreateJob, after}

\item{httpTarget}{A HTTP target object \link{HttpTarget}}

\item{description}{Optionally caller-specified in CreateJob or}

\item{projectId}{The GCP project to run within}
}
\description{
Creates a Cloud Scheduler job.
}
\examples{

\dontrun{
cr_init()
cr_schedule("* * * * *", name="test", httpTarget = HttpTarget(uri="https://code.markedmondson.me"))
}
}
\seealso{
\href{https://cloud.google.com/scheduler/}{Google Documentation}

Other Job functions: \code{\link{Job}},
  \code{\link{projects.locations.jobs.create}},
  \code{\link{projects.locations.jobs.patch}}
}
\concept{Job functions}
