% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_auth.R
\name{cr_setup_auth}
\alias{cr_setup_auth}
\title{Create a service account for googleCloudRunner}
\usage{
cr_setup_auth(
  email = Sys.getenv("GARGLE_EMAIL"),
  file = "googlecloudrunner-auth-key.json",
  session_user = NULL
)
}
\arguments{
\item{email}{What email to open OAuth2 with}

\item{file}{Where to save the authentication file}

\item{session_user}{1 for user level, 2 for project level, leave \code{NULL} to be prompted}
}
\value{
TRUE if the file is ready to be setup by \link{cr_setup}, FALSE if need to stop
}
\description{
This will use your Google OAuth2 user to create a suitable service account
}
\seealso{
Other setup functions: 
\code{\link{cr_setup_service}()},
\code{\link{cr_setup_test}()},
\code{\link{cr_setup}()}
}
\concept{setup functions}
