% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_tests.R
\name{cr_setup_test}
\alias{cr_setup_test}
\title{Run tests over your setup}
\usage{
cr_setup_test(
  option = c("menu", "all", "docker", "plumber", "r_script", "r_schedule")
)
}
\arguments{
\item{option}{Default will use an interactive menu, select other option to run that test without a menu}
}
\description{
This allows you to check if your setup works - run \link{cr_setup} first.
}
\examples{
\dontrun{
# start the menu for interactive use
cr_setup_test()

# skip menu and run all tests
cr_setup_test("all")

# run just the plumber deployment test
cr_setup_test("plumber")
}
}
\seealso{
Other setup functions: 
\code{\link{cr_setup_auth}()},
\code{\link{cr_setup_service}()},
\code{\link{cr_setup}()}
}
\concept{setup functions}
