% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleCloudRunner.R
\docType{package}
\name{googleCloudRunner}
\alias{googleCloudRunner}
\title{Launch R scripts into the Google Cloud via Cloud Build, Cloud Run and Cloud Scheduler}
\description{
See website for more details: \url{https://code.markedmondson.me/googleCloudRunner/}
}
