% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ListLocationsResponse}
\alias{ListLocationsResponse}
\title{ListLocationsResponse Object}
\usage{
ListLocationsResponse(locations = NULL, nextPageToken = NULL)
}
\arguments{
\item{locations}{A list of locations that matches the specified filter in the request}

\item{nextPageToken}{The standard List next-page token}
}
\value{
ListLocationsResponse object
}
\description{
ListLocationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for Locations.ListLocations.
}
\concept{ListLocationsResponse functions}
