% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_run.R, R/objs_schedule.R
\name{Location.metadata}
\alias{Location.metadata}
\title{Location.metadata Object}
\usage{
Location.metadata()

Location.metadata()
}
\value{
Location.metadata object

Location.metadata object
}
\description{
Location.metadata Object

Location.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Service-specific metadata. For example the available capacity at the givenlocation.

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Service-specific metadata. For example the available capacity at the givenlocation.
}
\seealso{
Other Location functions: \code{\link{Location.labels}},
  \code{\link{Location}}

Other Location functions: \code{\link{Location.labels}},
  \code{\link{Location}}
}
\concept{Location functions}
