% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudscheduler.R
\name{PubsubTarget}
\alias{PubsubTarget}
\title{Pubsub Target Object (Cloud Scheduler)}
\usage{
PubsubTarget(topicName = NULL, data = NULL, attributes = NULL)
}
\arguments{
\item{topicName}{The name of the Cloud Pub/Sub topic to which messages will be published when a job is delivered.}

\item{data}{The message payload for PubsubMessage. An R object that will be turned into JSON via [jsonlite] and then base64 encoded into the PubSub format.}

\item{attributes}{Attributes for PubsubMessage.}
}
\value{
PubsubTarget object
}
\description{
Pubsub Target Object (Cloud Scheduler)
}
\details{
Pub/Sub target. The job will be delivered by publishing a message to the given Pub/Sub topic.
}
\seealso{
Other Cloud Scheduler functions: 
\code{\link{HttpTarget}()},
\code{\link{Job}()},
\code{\link{cr_build_schedule_http}()},
\code{\link{cr_run_schedule_http}()},
\code{\link{cr_schedule_delete}()},
\code{\link{cr_schedule_get}()},
\code{\link{cr_schedule_list}()},
\code{\link{cr_schedule_pause}()},
\code{\link{cr_schedule_run}()}
}
\concept{Cloud Scheduler functions}
