% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{RepoSource}
\alias{RepoSource}
\title{RepoSource Object}
\usage{
RepoSource(tagName = NULL, projectId = NULL, repoName = NULL,
  commitSha = NULL, branchName = NULL, dir = NULL)
}
\arguments{
\item{tagName}{Regex matching tags to build}

\item{projectId}{ID of the project that owns the Cloud Source Repository}

\item{repoName}{Name of the Cloud Source Repository}

\item{commitSha}{Explicit commit SHA to build}

\item{branchName}{Regex matching branches to build}

\item{dir}{Directory, relative to the source root, in which to run the build}
}
\value{
RepoSource object
}
\description{
RepoSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Location of the source in a Google Cloud Source Repository.
}
\seealso{
Other RepoSource functions: \code{\link{projects.triggers.run}}
}
\concept{RepoSource functions}
