% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_build.R
\name{Source}
\alias{Source}
\title{Source Object}
\usage{
Source(repoSource = NULL, storageSource = NULL)
}
\arguments{
\item{repoSource}{If provided, get the source from this location in a Cloud Source}

\item{storageSource}{If provided, get the source from this location in Google Cloud Storage}
}
\value{
Source object
}
\description{
Source Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Location of the source in a supported storage service.
}
\concept{Source functions}
