% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Status}
\alias{Status}
\title{Status Object}
\usage{
Status(code = NULL, message = NULL, status = NULL, details = NULL,
  reason = NULL, metadata = NULL)
}
\arguments{
\item{code}{Suggested HTTP return code for this status, 0 if not set}

\item{message}{A human-readable description of the status of this operation}

\item{status}{Status of the operation}

\item{details}{Extended data associated with the reason}

\item{reason}{A machine-readable description of why this operation is in the}

\item{metadata}{Standard list metadata}
}
\value{
Status object
}
\description{
Status Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Status is a return value for calls that don't return other objects
}
\concept{Status functions}
