% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild.R
\name{StorageSource}
\alias{StorageSource}
\title{StorageSource Object}
\usage{
StorageSource(bucket = NULL, object = NULL, generation = NULL)
}
\arguments{
\item{bucket}{Google Cloud Storage bucket containing the source}

\item{object}{Google Cloud Storage object containing the source. This object must be a gzipped archive file (.tar.gz) containing source to build.}

\item{generation}{Google Cloud Storage generation for the object.  If the generation is omitted, the latest generation will be used.}
}
\value{
StorageSource object
}
\description{
StorageSource Object
}
\details{
Location of the source in an archive file in Google Cloud Storage.
}
\examples{

\dontrun{


my_gcs_source <- Source(storageSource=StorageSource("gs://my-bucket","my_code.tar.gz"))

build1 <- cr_build("cloudbuild.yaml", source = my_gcs_source)

storage <- cr_build_upload_gcs("my_folder")
my_gcs_source2 <- Source(storageSource=storage)

build2 <- cr_build("cloudbuild.yaml", source = my_gcs_source2)

}
}
\concept{StorageSource functions}
