% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Volume}
\alias{Volume}
\title{Volume Object}
\usage{
Volume(name = NULL, secret = NULL, configMap = NULL)
}
\arguments{
\item{name}{Cloud Run fully managed: not supported}

\item{secret}{Cloud Run fully managed: not supported}

\item{configMap}{Cloud Run fully managed: not supported}
}
\value{
Volume object
}
\description{
Volume Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Run fully managed: not supportedCloud Run for Anthos: supportedVolume represents a named volume in a container.
}
\concept{Volume functions}
