% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{Yaml}
\alias{Yaml}
\title{Helper to create yaml files}
\usage{
Yaml(...)
}
\arguments{
\item{...}{steps in the yaml object}
}
\description{
Helper to create yaml files
}
\examples{

Yaml(steps = c(
      cr_buildstep("docker", "version"),
      cr_buildstep("gcloud", "version")),
    images = "gcr.io/my-project/my-image")
}
\seealso{
Other Cloud Build functions, yaml functions: \code{\link{cr_build_write}},
  \code{\link{cr_buildstep}}
}
\concept{Cloud Build functions, yaml functions}
